categories and routes domain modeling and mock data
