+++ data-access-layer/mock/repositories/mock-goal-repository.ts
@@ -1,6 +1,6 @@
 import { Goal } from '../../../domain-import-only/Goal';
-import { Trick } from '../../../domain-import-only/Trick';
 import { IGoalReporistory } from '../../repository-interfaces/goal-repository';
+import { LEGACY_SOURCE_GOALS, legacy_goal_to_goal } from '../legacy-goals';
 import { MockTrickRepository } from './mock-trick-repository';
 
 export class MockGoalRepository implements IGoalReporistory {
@@ -16,7 +16,7 @@ export class MockGoalRepository implements IGoalReporistory {
     const trickRepository = new MockTrickRepository();
     const allTricks = await trickRepository.getAllTricks();
 
-    return legacyGoals.map((lg) => _legacy_goal_to_goal(lg, allTricks));
+    return legacyGoals.map((lg) => legacy_goal_to_goal(lg, allTricks));
   }
 
   /**
@@ -28,1469 +28,10 @@ export class MockGoalRepository implements IGoalReporistory {
     const trickRepository = new MockTrickRepository();
     const allTricks = await trickRepository.getAllTricks();
     for (const legacyGoal of LEGACY_SOURCE_GOALS) {
-      const goal = await _legacy_goal_to_goal(legacyGoal, allTricks);
+      const goal = await legacy_goal_to_goal(legacyGoal, allTricks);
       goals.push(goal);
     }
 
     return goals;
   }
 }
-
-/**
- *
- * @param legacyGoal a goal imported from legacy bingo teacher google sheet
- * @param allTricks a list of all tricks which may align with the trickIds from the google sheet - note their ID must be correct
- * @returns A correctly built modern Goal domain model
- */
-const _legacy_goal_to_goal = (legacyGoal: LegacyGoal, allTricks: Trick[]) => {
-  const tricks = legacyGoal.trickIds.map((trickId) => {
-    if (trickId !== '') {
-      const trick = allTricks.find((t) => t.id === trickId);
-      if (!trick) {
-        throw new Error(`Trick with id ${trickId} not found.`);
-      }
-      return trick;
-    }
-  });
-
-  const notes = legacyGoal.goalNotes.join('\n');
-
-  const goal = Goal.create({
-    name: legacyGoal.goalName,
-    notes: notes,
-    tricks: tricks,
-  });
-
-  return goal;
-};
-
-interface LegacyGoal {
-  goalName: string;
-  trickIds: string[];
-  goalNotes: string[];
-}
-
-/**
- * This is raw data from the existing bingo teacher google doc. It'll die eventually
- */
-const LEGACY_SOURCE_GOALS: LegacyGoal[] = [
-  {
-    goalName: '3 Lake Hylia Skulltulas',
-    trickIds: ['C34B7975'],
-    goalNotes: [
-      'Other two child skulls are in the soil patch in front of the lab with bugs',
-      'and on the fire arrow platform.  You can also get one as adult in the tree above water temple with HS.',
-    ],
-  },
-  {
-    goalName: 'All 4 Lon-Lon Ranch area Skulltulas',
-    trickIds: ['60531191'],
-    goalNotes: [''],
-  },
-  {
-    goalName: 'Beat the Water Temple',
-    trickIds: ['70DAAA62', 'DD4D7F1C', 'C3FBA8DC', '6DD78777'],
-    goalNotes: [''],
-  },
-  {
-    goalName: "Defeat all Lizalfos in Dodongo's Cavern",
-    trickIds: ['B69C4380'],
-    goalNotes: [
-      'Doom jump (preferably)',
-      'get bombs ONLY IF NEEDED.\nBacktrack from there (reverse flames). Top lizalfos',
-      'hookshot from bombflower to ladder.\nThen jump down and go to lizalfos behind the switch.',
-    ],
-  },
-  {
-    goalName: 'Map & Compass in Water Temple',
-    trickIds: ['771D2D81', 'D80E6EDA', '1E196CEA', 'D7357272'],
-    goalNotes: [''],
-  },
-  {
-    goalName: 'Water Medallion',
-    trickIds: [
-      '5CC02919',
-      'A849D697',
-      '70DAAA62',
-      'DD4D7F1C',
-      'C3FBA8DC',
-      '6DD78777',
-    ],
-    goalNotes: ['Beat Water or RBA bugs with broken sword on C-Right'],
-  },
-  {
-    goalName: 'Water Temple Boss Key',
-    trickIds: ['5CC02919', 'A849D697', 'A596CE56', '2BD6E687'],
-    goalNotes: ['RBA Bugs with Ice arrows on C-Right'],
-  },
-  {
-    goalName: '1 Skulltula from each Adult Dungeon',
-    trickIds: [
-      '9C338AB0',
-      '4657BEE7',
-      '8FAE21BA',
-      '11DF7000',
-      '1E196CEA',
-      '85D1C987',
-      'B4E48D8C',
-      '00D651EC',
-      'D7E0B9BE',
-    ],
-    goalNotes: [''],
-  },
-  {
-    goalName: '1 Skulltula from each Child Dungeon',
-    trickIds: ['B669724E', '13E1BC06', '933844BC', '33028619'],
-    goalNotes: [
-      'Jabu: If you just need skull',
-      'jump down and backtrack instead of talking to ruto.',
-    ],
-  },
-  {
-    goalName: '1 Unused Small Key in each Adult Dungeon',
-    trickIds: [
-      '9C338AB0',
-      '4657BEE7',
-      '8FAE21BA',
-      '1E196CEA',
-      'D7357272',
-      '4E0619E0',
-      'B4E48D8C',
-      'B090EE3A',
-      '3C1E3767',
-      'D7E0B9BE',
-    ],
-    goalNotes: ['can rba fake bomb bag for water keys'],
-  },
-  {
-    goalName: '10 Songs',
-    trickIds: ['A849D697', '5CC02919'],
-    goalNotes: [
-      'RBA Bugs with Odd Potion then Saw on C-Right',
-      'then collect SoS, Suns, and remaining warp songs.',
-    ],
-  },
-  {
-    goalName: '15 Different Skulltulas',
-    trickIds: [],
-    goalNotes: [''],
-  },
-  {
-    goalName: '2 Boss Keys',
-    trickIds: ['5CC02919', 'A596CE56', 'A17F4252'],
-    goalNotes: [
-      'RBA Elemental Arrows (Fire=Fire, Ice=Water, Light=Spirit)',
-      ' or collect from adult dungeons/ Ganons Tower',
-    ],
-  },
-  {
-    goalName: '3 Boots',
-    trickIds: ['D7E0B9BE', '827B2011', '95759E64'],
-    goalNotes: [''],
-  },
-  {
-    goalName: '3 Boss Keys',
-    trickIds: ['5CC02919', 'A596CE56', 'A17F4252'],
-    goalNotes: [
-      'RBA Elemental Arrows (Fire=Fire, Ice=Water, Light=Spirit)',
-      ' or collect from adult dungeons/ Ganons Tower',
-    ],
-  },
-  {
-    goalName: '3 Shields',
-    trickIds: ['B090EE3A'],
-    goalNotes: [''],
-  },
-  {
-    goalName: '3 Shields & 3 Boots',
-    trickIds: ['B090EE3A', 'D7E0B9BE', '827B2011', '95759E64'],
-    goalNotes: [''],
-  },
-  {
-    goalName: '3 Shields & 3 Tunics',
-    trickIds: ['B090EE3A', 'B0CA9C91', 'B1F50951'],
-    goalNotes: [''],
-  },
-  {
-    goalName: '3 Skulltulas in Water Temple',
-    trickIds: ['1E196CEA', 'D7357272', '85D1C987', 'C51AA108'],
-    goalNotes: [
-      'Route Video is split into 3 sections',
-      'one for if you have ZL',
-      'One for if you have Iron Boots',
-      'and one for if you just have a bunch of chus to hover.',
-    ],
-  },
-  {
-    goalName: '3 Songs',
-    trickIds: ['5CC02919', 'A849D697'],
-    goalNotes: [''],
-  },
-  {
-    goalName: '3 Swords & 3 Boots',
-    trickIds: ['D7E0B9BE', '827B2011', '95759E64'],
-    goalNotes: [''],
-  },
-  {
-    goalName: '3 Swords & 3 Shields',
-    trickIds: ['B090EE3A'],
-    goalNotes: [''],
-  },
-  {
-    goalName: '3 Swords & 3 Tunics',
-    trickIds: ['B0CA9C91', 'B1F50951'],
-    goalNotes: [''],
-  },
-  {
-    goalName: '3 Swords, Tunics, Boots, and Shields',
-    trickIds: ['B090EE3A', 'B0CA9C91', 'B1F50951'],
-    goalNotes: [''],
-  },
-  {
-    goalName: '3 Tunics',
-    trickIds: ['B0CA9C91', 'B1F50951'],
-    goalNotes: [''],
-  },
-  {
-    goalName: '3 Tunics & 3 Boots',
-    trickIds: ['B0CA9C91', 'D7E0B9BE', '827B2011', '95759E64', 'B1F50951'],
-    goalNotes: [''],
-  },
-  {
-    goalName: '3 Unused Keys in Gerudo Training Grounds',
-    trickIds: ['7B3EABC2', '2C5B71A5'],
-    goalNotes: [''],
-  },
-  {
-    goalName: '30 Deku Nuts',
-    trickIds: [],
-    goalNotes: [
-      'Buy Upgrade from Deku Scrub under bombable boulder in Lost woods near exit to Sacred Forest Medow. (4 way before the exit to SFM)',
-    ],
-  },
-  {
-    goalName: '37th Heart Piece (Child Fortress)',
-    trickIds: ['3625CB18', '9680FAC3'],
-    goalNotes: [''],
-  },
-  {
-    goalName: '4 Compasses',
-    trickIds: [],
-    goalNotes: ['BotW', 'DC', 'and Shadow are common'],
-  },
-  {
-    goalName: '4 Maps',
-    trickIds: ['5CC02919', 'A849D697'],
-    goalNotes: ['Shadow and DC are common'],
-  },
-  {
-    goalName: '4 Skulltulas in Shadow Temple',
-    trickIds: ['B6828BD7'],
-    goalNotes: ["Don't get Skulltula past boat", 'which requires ZL.'],
-  },
-  {
-    goalName: '4 Songs',
-    trickIds: ['5CC02919', 'A849D697'],
-    goalNotes: ['RBA Bugs with Saw on C-Right'],
-  },
-  {
-    goalName: '4 Unused Keys in Forest Temple',
-    trickIds: ['9C338AB0', '4B71852C', '06950182', '04B44A63'],
-    goalNotes: [
-      'Mega jump gets you to floormaster room with key. Last key is well key.',
-    ],
-  },
-  {
-    goalName: '4 Unused Keys in Gerudo Training Grounds',
-    trickIds: ['7B3EABC2', '2C5B71A5'],
-    goalNotes: [''],
-  },
-  {
-    goalName: '5 Compasses',
-    trickIds: [],
-    goalNotes: ['BotW', 'DC', 'and Shadow are common'],
-  },
-  {
-    goalName: '5 Hearts',
-    trickIds: ['F09CEAC1'],
-    goalNotes: [''],
-  },
-  {
-    goalName: '5 Magic Beans',
-    trickIds: ['D14D73DC'],
-    goalNotes: [''],
-  },
-  {
-    goalName: '5 Maps',
-    trickIds: ['5CC02919', 'A849D697'],
-    goalNotes: [''],
-  },
-  {
-    goalName: '5 Unused Keys in Gerudo Training Grounds',
-    trickIds: ['7B3EABC2', '2C5B71A5'],
-    goalNotes: [''],
-  },
-  {
-    goalName: '5 Zora area HPs',
-    trickIds: ['4674313E', 'B5BDCB36'],
-    goalNotes: [
-      "1 HP is child only and requires lighting all the torches in Zora's Domain with Deku stick",
-      "2 are in Zora's river and can be quickly gotten as adult with Hover Boots or megaside hops",
-      'last 2 pieces are in Adult Zoras fountain one is freestanding on the frozen platforms',
-      'one is at the bottom of the water and requires Iron boots.  The HP in ice cavern does NOT count.',
-    ],
-  },
-  {
-    goalName: '500 Rupees',
-    trickIds: ['5CC02919', 'A849D697'],
-    goalNotes: [
-      'RBA with Pocket Cucco on C-Right',
-      'Collect 10 Skull reward',
-      'collect 500 Rupees',
-    ],
-  },
-  {
-    goalName: '6 Compasses',
-    trickIds: [],
-    goalNotes: ['BotW', 'DC', 'and Shadow are common'],
-  },
-  {
-    goalName: '6 Hearts',
-    trickIds: ['F09CEAC1'],
-    goalNotes: [''],
-  },
-  {
-    goalName: '6 Maps',
-    trickIds: ['5CC02919', 'A849D697'],
-    goalNotes: [''],
-  },
-  {
-    goalName: '6 Songs',
-    trickIds: ['5CC02919', 'A849D697'],
-    goalNotes: [
-      'RBA Bugs with Odd Potion then Saw on C-Right',
-      'then collect one additional song',
-    ],
-  },
-  {
-    goalName: '6 Unused Keys in Gerudo Training Grounds',
-    trickIds: ['7B3EABC2', '2C5B71A5'],
-    goalNotes: [''],
-  },
-  {
-    goalName: '7 Compasses',
-    trickIds: [],
-    goalNotes: ['There are 10 compasses in the game'],
-  },
-  {
-    goalName: '7 Different Unused Keys in Gerudo Training Grounds',
-    trickIds: ['7B3EABC2', '2C5B71A5'],
-    goalNotes: [''],
-  },
-  {
-    goalName: '7 Hearts',
-    trickIds: ['F09CEAC1'],
-    goalNotes: ['Just get additional HPs along your route'],
-  },
-  {
-    goalName: '7 Magic Beans',
-    trickIds: ['D14D73DC'],
-    goalNotes: [''],
-  },
-  {
-    goalName: '7 Maps',
-    trickIds: ['5CC02919', 'A849D697'],
-    goalNotes: [''],
-  },
-  {
-    goalName: '7 Songs',
-    trickIds: ['5CC02919', 'A849D697'],
-    goalNotes: [
-      'RBA Bugs with Odd Potion then Saw on C-Right',
-      'then collect two additional songs.',
-    ],
-  },
-  {
-    goalName: '8 Different Unused Keys in Gerudo Training Grounds',
-    trickIds: ['7B3EABC2', '2C5B71A5'],
-    goalNotes: [''],
-  },
-  {
-    goalName: '8 Hearts (no duping)',
-    trickIds: [],
-    goalNotes: [''],
-  },
-  {
-    goalName: '8 Songs',
-    trickIds: ['5CC02919', 'A849D697'],
-    goalNotes: [
-      'RBA Bugs with Odd Potion then Saw on C-Right',
-      'then collect three additional songs.',
-    ],
-  },
-  {
-    goalName: '9 Hearts (no duping)',
-    trickIds: [],
-    goalNotes: [''],
-  },
-  {
-    goalName: '9 Songs',
-    trickIds: ['5CC02919', 'A849D697'],
-    goalNotes: [
-      'RBA Bugs with Odd Potion then Saw on C-Right',
-      'then collect 4 additional songs on your route.',
-    ],
-  },
-  {
-    goalName: 'All 3 Elemental Arrows',
-    trickIds: [
-      '3B9CB785',
-      '8AB11EE5',
-      '5EE45339',
-      'A596CE56',
-      '8C37492D',
-      'A17F4252',
-      '1E196CEA',
-      '70DAAA62',
-      'DD4D7F1C',
-      'C3FBA8DC',
-    ],
-    goalNotes: ['Common to RBA Broken Sword for LACS'],
-  },
-  {
-    goalName: 'All 3 Kokiri Forest area Skulltulas',
-    trickIds: [],
-    goalNotes: [
-      'Two Child skulls are behind twins house and soil patch behind shop',
-      'Adult skultula is on top of house to the right of the path to Deku tree',
-    ],
-  },
-  {
-    goalName: 'All 3 Skulltulas in Bottom of the Well',
-    trickIds: ['3126A906', '061F7FDE', '06B33728'],
-    goalNotes: [
-      'You do not have to vineclip for Chus.  On 1.0 you can do blank A and on 1.2 you can do ocarina dive. The rest of the route is the same.',
-    ],
-  },
-  {
-    goalName: 'All 3 Skulltulas in Ice Cavern',
-    trickIds: ['0F804DD9', '827B2011'],
-    goalNotes: [''],
-  },
-  {
-    goalName: 'All 4 Gerudo Valley area Skulltulas',
-    trickIds: ['B0C59F81'],
-    goalNotes: [''],
-  },
-  {
-    goalName: 'All 4 Lost Woods area Skulltulas',
-    trickIds: ['81D84A88'],
-    goalNotes: [''],
-  },
-  {
-    goalName: 'All 4 Market area Skulltulas',
-    trickIds: ['048B036A'],
-    goalNotes: ['Song of Storms grotto requires explosives or blue fire'],
-  },
-  {
-    goalName: 'All 4 Skulltulas in Deku Tree',
-    trickIds: ['13E1BC06'],
-    goalNotes: [''],
-  },
-  {
-    goalName: 'All 4 Skulltulas in Jabu-Jabu',
-    trickIds: ['33028619', '933844BC'],
-    goalNotes: [''],
-  },
-  {
-    goalName: 'All 4 Wasteland/ Colossus area Skulltulas',
-    trickIds: ['3625CB18', 'E9E1A81E', 'D3D2189E', '375E2213'],
-    goalNotes: [''],
-  },
-  {
-    goalName: 'All 5 Lake Hylia Skulltulas',
-    trickIds: ['C34B7975'],
-    goalNotes: [
-      'Other two child skulls are in the soil patch in front of the lab with bugs',
-      'and on the fire arrow platform.  You can also get one as adult in the tree above water temple with HS. Skull in Lab requires ironboots.',
-    ],
-  },
-  {
-    goalName: "All 5 Skulltulas in Dodongo's Cavern",
-    trickIds: ['BA12A4D9'],
-    goalNotes: [''],
-  },
-  {
-    goalName: 'All 5 Skulltulas in Fire Temple',
-    trickIds: ['6C288E6E', '8FAE21BA', '6CA81599', '11DF7000', '89A994DA'],
-    goalNotes: [''],
-  },
-  {
-    goalName: 'All 5 Skulltulas in Forest Temple',
-    trickIds: ['9619C13C', '9C338AB0', '815D2CE3'],
-    goalNotes: [''],
-  },
-  {
-    goalName: 'All 5 Skulltulas in Shadow Temple',
-    trickIds: ['B4E48D8C', 'B6828BD7', '00D651EC', '5BAD1B80'],
-    goalNotes: [''],
-  },
-  {
-    goalName: 'All 5 Skulltulas in Spirit Temple',
-    trickIds: ['614800B4', '42513477', 'D7E0B9BE', '3C1E3767'],
-    goalNotes: [''],
-  },
-  {
-    goalName: 'All 5 Skulltulas in Water Temple',
-    trickIds: ['1E196CEA', 'D7357272', '7CB759AA', '120D7D11'],
-    goalNotes: [''],
-  },
-  {
-    goalName: 'All 8 Death Mountain area Skulltulas',
-    trickIds: [],
-    goalNotes: [''],
-  },
-  {
-    goalName: 'All 8 Kakariko area Skulltulas',
-    trickIds: ['3E3DECA3'],
-    goalNotes: [
-      "Once you're on the seam you can kill the GS with a bombchu then drop onto it by pulling sword.  Other child skulls are behind bazaar",
-      'on ladder',
-      'on achery construction site',
-      'behind skull house',
-      'graveyard soil patch',
-      "bonk into tree. Adult GS is on top of impa's house",
-      'can be gotten with HS.',
-    ],
-  },
-  {
-    goalName: "All 8 Zora's Domain area Skulltulas",
-    trickIds: ['366B4776', '089F7B22'],
-    goalNotes: [''],
-  },
-  {
-    goalName: "Beat Dodongo's Cavern",
-    trickIds: ['B69C4380', '6C2A9F69'],
-    goalNotes: [''],
-  },
-  {
-    goalName: "Beat Jabu-Jabu's Belly",
-    trickIds: ['933844BC'],
-    goalNotes: [''],
-  },
-  {
-    goalName: 'Beat the Deku Tree',
-    trickIds: ['B669724E', '7958F88E', '2B40FB6F'],
-    goalNotes: [''],
-  },
-  {
-    goalName: 'Beat the Fire Temple',
-    trickIds: [
-      '4657BEE7',
-      'E91970CC',
-      '6CA81599',
-      '8FAE21BA',
-      '76C797C9',
-      'C036348D',
-      'E7A8AC29',
-    ],
-    goalNotes: [''],
-  },
-  {
-    goalName: 'Beat the Forest Temple',
-    trickIds: ['9C338AB0', 'ED22ECBC', 'A9F572CC', '0A1DD7ED'],
-    goalNotes: [''],
-  },
-  {
-    goalName: 'Beat the Shadow Temple',
-    trickIds: ['B4E48D8C', '87D42917', 'DB6FF011', 'A0C2A8B5'],
-    goalNotes: [''],
-  },
-  {
-    goalName: 'Beat the Spirit Temple',
-    trickIds: ['B090EE3A', 'FF124B82', 'E33B98C0', 'B1F254D7', 'FB73AEF8'],
-    goalNotes: [''],
-  },
-  {
-    goalName: 'Blue Fire',
-    trickIds: [],
-    goalNotes: [
-      'Get Blue fire from Ice Cavern or RBA Pocket Cucco for adult wallet',
-      'then turn in 10 GS reward for giant wallet and by blue fire.',
-    ],
-  },
-  {
-    goalName: 'Blue Gauntlets',
-    trickIds: ['B090EE3A', '5CC02919', 'A849D697'],
-    goalNotes: [
-      'Have Silver Guantlets then RBA with Pocket Cucco on C-Right',
-      'afterwards you cannot RBA Cojiro or you will lose your Gauntlet color and the goal.',
-    ],
-  },
-  {
-    goalName: 'Blue Potion',
-    trickIds: [],
-    goalNotes: [
-      'Do adult trade all the way through getting odd potion then you can buy blue potion if youhave an empty bottle.',
-    ],
-  },
-  {
-    goalName: 'Bolero of Fire',
-    trickIds: ['4657BEE7'],
-    goalNotes: [''],
-  },
-  {
-    goalName: 'Bomb Bag (30)',
-    trickIds: [],
-    goalNotes: [''],
-  },
-  {
-    goalName: 'Boomerang',
-    trickIds: [],
-    goalNotes: [''],
-  },
-  {
-    goalName: 'Both Gerudo Valley HPs',
-    trickIds: [],
-    goalNotes: [''],
-  },
-  {
-    goalName: "Both Gerudo's Fortress area Skulltulas",
-    trickIds: [],
-    goalNotes: [''],
-  },
-  {
-    goalName: 'Both HPs in Death Mountain Crater',
-    trickIds: ['F4B0B104'],
-    goalNotes: ['The second HP is on the climb-able wall from the top of DMC'],
-  },
-  {
-    goalName: 'Both HPs in Lost Woods',
-    trickIds: ['997CEAC9', '7FD1B460'],
-    goalNotes: [
-      'Get ZL as Child',
-      "Get Saria's Song",
-      "Play Saria's for Skull kid in Lost Woods",
-      'Play Ocarina memory game with Skull Kids in Lost Woods beside Slingshot minigame.',
-    ],
-  },
-  {
-    goalName: 'Both Hyrule Field area Skulltulas',
-    trickIds: ['6F1BB8D3'],
-    goalNotes: [
-      'The second skull is in a bomable grotto by the far tree to the right as you exit Kak.',
-    ],
-  },
-  {
-    goalName: 'Bottled Fairy',
-    trickIds: [],
-    goalNotes: [
-      'Get a fairy in a bottle (must stay in inventory',
-      'best to dupe over uneeded item)',
-    ],
-  },
-  {
-    goalName: 'Bronze Gauntlets',
-    trickIds: ['8263CAC2', '5CC02919', 'A849D697', 'FA10A855'],
-    goalNotes: [
-      'RBA Bugs or fish with Pocket Cucco on C-Right after getting gold gauntlets.  After this you cannot RBA Cojiro or you will lost your gauntlets and the goal.',
-    ],
-  },
-  {
-    goalName: 'Bullet Bag (40)',
-    trickIds: [],
-    goalNotes: [
-      'Get fairy sling shot as child and do lost woods target game. Do not RBA with Pocket Cucco on C-right after getting this or you will lose your seedbag and the goal.',
-    ],
-  },
-  {
-    goalName: 'Bullet Bag (50)',
-    trickIds: [],
-    goalNotes: [
-      'Leave forest without getting Fairy Slingshot and do child shooting gallery. Do not RBA with Pocket Cucco on C-right after getting this or you will lose your seedbag and the goal.',
-    ],
-  },
-  {
-    goalName: 'Cow in House',
-    trickIds: ['2ECE35A9'],
-    goalNotes: [
-      "Steal Epona the go back in to Lon Lon Ranch on Epona and do Malon's minigame.",
-    ],
-  },
-  {
-    goalName: 'Defeat 4 Different Iron Knuckles',
-    trickIds: [
-      'FA10A855',
-      'EBBD3E69',
-      'A87720FA',
-      '452A0F7F',
-      'B090EE3A',
-      'B1F254D7',
-      '1AD2FFF3',
-    ],
-    goalNotes: [
-      'Spirit hover to SIlver chest hand go into temple and beat Knuckle',
-      'leave back to silver gaunts chest and super slide hover acrossed',
-      'enter temple and beat other Knuckle',
-      "get into Gannon's tower and beat double knuckles.",
-    ],
-  },
-  {
-    goalName: 'Defeat a Skull Kid',
-    trickIds: ['1660C305'],
-    goalNotes: [
-      'As adult',
-      "skull kid only appears before getting pocket egg or after getting poacher's saw.  As child",
-      "you must play Saria's song for him before he can be attacked",
-    ],
-  },
-  {
-    goalName: 'Defeat a White Wolfos',
-    trickIds: ['5EE45339'],
-    goalNotes: ['Ice Cavern or GTG'],
-  },
-  {
-    goalName: 'Defeat Amy (Green Poe)',
-    trickIds: ['9C338AB0', 'ACBB4ECB'],
-    goalNotes: [''],
-  },
-  {
-    goalName: 'Defeat Barinade',
-    trickIds: ['5FCDBB54', '933844BC'],
-    goalNotes: [''],
-  },
-  {
-    goalName: 'Defeat Big Octo',
-    trickIds: ['32F255A9', '4674313E', 'B5BDCB36', '933844BC'],
-    goalNotes: ['can be done without boomerang by using deku nuts'],
-  },
-  {
-    goalName: 'Defeat Bongo-Bongo',
-    trickIds: ['B4E48D8C', '87D42917', 'DB6FF011', 'A0C2A8B5'],
-    goalNotes: [''],
-  },
-  {
-    goalName: 'Defeat both Flare Dancers',
-    trickIds: ['E91970CC', '6CA81599'],
-    goalNotes: [''],
-  },
-  {
-    goalName: 'Defeat Dark Link',
-    trickIds: ['380D122E', '1E196CEA', 'D7357272'],
-    goalNotes: [''],
-  },
-  {
-    goalName: 'Defeat King Dodongo',
-    trickIds: ['B69C4380', '6C2A9F69'],
-    goalNotes: [''],
-  },
-  {
-    goalName: 'Defeat Meg (purple Poe)',
-    trickIds: ['9C338AB0', '04B44A63', 'ACBB4ECB'],
-    goalNotes: [
-      'Follow forest route to bow',
-      'beat red and blue poes',
-      'save warp to entrance of temple',
-      'hover boost to green poe',
-      'the leave green poe room to start purple poe fight.',
-    ],
-  },
-  {
-    goalName: 'Defeat Morpha',
-    trickIds: [
-      '1E196CEA',
-      'D7357272',
-      '70DAAA62',
-      '6DD78777',
-      'DD4D7F1C',
-      'C3FBA8DC',
-    ],
-    goalNotes: [''],
-  },
-  {
-    goalName: 'Nabooru Route',
-    trickIds: ['27D58B62'],
-    goalNotes: ['Just do the low height groudn clip bk skip'],
-  },
-  {
-    goalName: 'Defeat Phantom Ganon',
-    trickIds: ['9C338AB0', '03615B32', 'ED22ECBC', 'A9F572CC', '0A1DD7ED'],
-    goalNotes: [''],
-  },
-  {
-    goalName: 'Defeat Queen Gohma',
-    trickIds: ['B669724E', '7958F88E', '2B40FB6F'],
-    goalNotes: [
-      'Make sure to equip Kokiri boots after cliping',
-      'and have some nuts for Gohma',
-    ],
-  },
-  {
-    goalName: 'Defeat Twinrova',
-    trickIds: ['B090EE3A', 'FF124B82', 'E33B98C0', 'B1F254D7', 'FB73AEF8'],
-    goalNotes: [''],
-  },
-  {
-    goalName: 'Defeat Volvagia',
-    trickIds: [
-      '4657BEE7',
-      'E91970CC',
-      '6CA81599',
-      '8FAE21BA',
-      '76C797C9',
-      'C036348D',
-      'E7A8AC29',
-    ],
-    goalNotes: [''],
-  },
-  {
-    goalName: "Din's Fire",
-    trickIds: ['997CEAC9', '7FD1B460'],
-    goalNotes: [
-      'Get ZL and magic as child then go to fairy fountain in Hyrule castle and get dins',
-    ],
-  },
-  {
-    goalName: 'Double Magic',
-    trickIds: ['62527303'],
-    goalNotes: ['Requires ZL and Magic'],
-  },
-  {
-    goalName: "Epona's Song",
-    trickIds: ['997CEAC9'],
-    goalNotes: [
-      'Get egg from Malon wake up Talon go to Lon Lon Ranch and get Eponas',
-    ],
-  },
-  {
-    goalName: 'Exactly 20 Deku Sticks',
-    trickIds: ['5CC02919', 'A849D697'],
-    goalNotes: [
-      'Buy first Upgrade from Business scrub in Lost Woods area below Kokiri Bridge',
-      'or RBA bugs with Pocket Egg on C-Right to get 30 capacity.',
-    ],
-  },
-  {
-    goalName: 'Exactly 30 Deku Sticks',
-    trickIds: ['5CC02919', 'A849D697'],
-    goalNotes: [
-      ' RBA bugs with Pocket Egg on C-Right to get 30 capacity. RBA bugs with Fairy on C-Right to get 29 then buy/get one.',
-    ],
-  },
-  {
-    goalName: 'Fairy Bow',
-    trickIds: ['9C338AB0', '04B44A63'],
-    goalNotes: [''],
-  },
-  {
-    goalName: 'Fairy Slingshot',
-    trickIds: ['B669724E'],
-    goalNotes: ['Once you are in Deku just go get slingshot'],
-  },
-  {
-    goalName: "Farore's Wind",
-    trickIds: ['5CC02919', 'A849D697'],
-    goalNotes: [
-      'RBA Bugs on B with Poachers Saw on C-Right for ZL',
-      'Get magic at the top of DMT',
-      "then go get Farore's",
-    ],
-  },
-  {
-    goalName: 'Fill all 4 Bottle Slots',
-    trickIds: ['5CC02919', 'A849D697', '4674313E'],
-    goalNotes: [
-      'Do child cuccos in Kakariko for first bottle',
-      "get Ruto's letter for second bottle slot",
-      'RBA with empty bottle on C-Right for 3rd bottle Slot',
-      "RBA with red Potion on C-Right for fourth bottle slot (If you route includes magics or going to Spirit temple with deep RBA then you can skip Ruto's letter and RBA with Nayru's Love on C-Right for second bottle slot.)",
-    ],
-  },
-  {
-    goalName: 'Fire Arrows',
-    trickIds: ['8C37492D', 'A17F4252'],
-    goalNotes: [''],
-  },
-  {
-    goalName: 'Fire Medallion',
-    trickIds: [
-      '4657BEE7',
-      'E91970CC',
-      '6CA81599',
-      '8FAE21BA',
-      '76C797C9',
-      'C036348D',
-      'E7A8AC29',
-      '815D2CE3',
-    ],
-    goalNotes: [''],
-  },
-  {
-    goalName: 'Fire Temple Boss Key',
-    trickIds: [
-      '1E196CEA',
-      'D7357272',
-      '70DAAA62',
-      '6DD78777',
-      'DD4D7F1C',
-      'C3FBA8DC',
-      '8C37492D',
-      'A17F4252',
-    ],
-    goalNotes: ['RBA Bugs or Fish with fire arros on C-right'],
-  },
-  {
-    goalName: 'Forest Medallion',
-    trickIds: [
-      '5CC02919',
-      'A849D697',
-      '9C338AB0',
-      'ED22ECBC',
-      'A9F572CC',
-      '0A1DD7ED',
-    ],
-    goalNotes: ['Beat Forest or RBA bugs with broken sword on C-Right'],
-  },
-  {
-    goalName: 'Forest Temple Boss Key',
-    trickIds: ['04B44A63', '5CC02919', 'A849D697', '2BD6E687', '7B3EABC2'],
-    goalNotes: [
-      'After you get bow go back up to eyeball switch and flip BK room to get BK',
-      'with Ice Arrows you have to first RBA Cojiro and dupe a bottle over the bow slot.',
-    ],
-  },
-  {
-    goalName: 'Free all 9 Gorons in Fire Temple',
-    trickIds: ['4657BEE7', '8FAE21BA', '6CA81599'],
-    goalNotes: [
-      'First Goron in the left room were Darunia is',
-      '2-3 are in the first level lava room one to the left as you enter and one to the right',
-      '4 is in the block push room with the crystal switch (Clip to this Goron)',
-      '5-7 are in the rolling boulder maze one up top two on the bottom',
-      '8 is after you get hammer drop down to the right and jumpslash the hammer switch through the Song of Time block',
-      '9 is past the hammerable totem in the first room of the temple (This requires a key make sure to grab keys from Gorons 1-5)',
-    ],
-  },
-  {
-    goalName: "Frog's HP",
-    trickIds: ['9C338AB0', 'ED22ECBC', 'A9F572CC'],
-    goalNotes: [
-      "Requires back in time so you must beat Forest Temple. Get Song of Storms and play it for the frogs in Zora's River as child.",
-    ],
-  },
-  {
-    goalName: "Ganon's Castle Boss Key",
-    trickIds: ['FA10A855', 'EBBD3E69', 'A87720FA', '452A0F7F', '1AD2FFF3'],
-    goalNotes: [''],
-  },
-  {
-    goalName: "Gerudo's Card",
-    trickIds: ['3B9CB785', '8AB11EE5'],
-    goalNotes: ['Just beat the gerudo guards and free the carpenters.'],
-  },
-  {
-    goalName: 'Get Bombchu chest in Spirit Temple',
-    trickIds: ['3B9CB785', '8AB11EE5', '375E2213', 'D7E0B9BE', '3C1E3767'],
-    goalNotes: [
-      'Chest is in room with Lizalfos',
-      'bombable wall and sun target on floor.  Hit the crystal switch to drop the bombchu chest.',
-    ],
-  },
-  {
-    goalName: 'Get to the end of Fire Trial',
-    trickIds: ['FA10A855', '1BFB5B64'],
-    goalNotes: ['Much easier with gold gauntlet strength'],
-  },
-  {
-    goalName: 'Get to the end of Forest Trial',
-    trickIds: ['FA10A855', 'E67D7D5A'],
-    goalNotes: [
-      'Requires Magic and a fire source.  Either get ZL',
-      "Mgic and Din's as child",
-      'then HS to top torch and light with Dins',
-      'or RBA cojiro for quiver',
-      'RBA Saw for ZL get magic and fire arrows.',
-    ],
-  },
-  {
-    goalName: 'Get to the end of Light Trial',
-    trickIds: ['FA10A855', 'FB51F631'],
-    goalNotes: [
-      'Requires ZL',
-      'RBA Bugs with Saw on C-Right',
-      'You can also unload the block by entering and exiting Spirit trial many times.',
-    ],
-  },
-  {
-    goalName: 'Get to the end of Shadow Trial',
-    trickIds: ['FA10A855', '8263CAC2'],
-    goalNotes: [''],
-  },
-  {
-    goalName: 'Get to the end of Spirit Trial',
-    trickIds: ['FA10A855'],
-    goalNotes: [
-      'Spirit trial is easy with HS and Chus',
-      "in a bingo there isn't much reason to be here without those two things.  You can shot the web with just bow from behind the torch in the torch slug room.",
-    ],
-  },
-  {
-    goalName: 'Get to the end of Water Trial',
-    trickIds: ['FA10A855', '9E248370'],
-    goalNotes: [''],
-  },
-  {
-    goalName: "Giant's Knife",
-    trickIds: ['5CC02919', 'A849D697'],
-    goalNotes: [
-      'RBA with Pocket Cucco on C-Right for adults wallet',
-      'get 200 rupees',
-      'buy Giants knife in Goron City.',
-    ],
-  },
-  {
-    goalName: "Giant's Wallet",
-    trickIds: ['5CC02919', 'A849D697'],
-    goalNotes: ['RBA with Pocket Cucco on C-Right', 'Collect 10 Skulls reward'],
-  },
-  {
-    goalName: 'Golden Gauntlets',
-    trickIds: ['FA10A855', '8263CAC2'],
-    goalNotes: [''],
-  },
-  {
-    goalName: 'Goron Bracelet',
-    trickIds: ['997CEAC9', '7FD1B460'],
-    goalNotes: [
-      'Get ZL as Child',
-      "Get Saria's go to Goron City and play Saria's for Darunia. You cannot RBA Cojio after this or you will lose your Goron Bracelt and the goal.",
-    ],
-  },
-  {
-    goalName: 'Goron Tunic',
-    trickIds: ['39077DAC', '1660C305'],
-    goalNotes: [
-      'Stop \'Link of the Gorons" in adult Goron City talk through both his text options or buy for 200 rupees via rba wallet. Killing Skull Kid is an easy way to get 200 Rupees',
-    ],
-  },
-  {
-    goalName: 'Green Gauntlets',
-    trickIds: ['997CEAC9', '7FD1B460', '5CC02919', 'A849D697'],
-    goalNotes: [
-      'Get Goron Bracelet then RBA with Pocket Cucco on C-Right.  Afterwards you cannot RBA Cojiro or you will lose your Gauntlet color and the goal',
-    ],
-  },
-  {
-    goalName: 'Ice Arrows',
-    trickIds: ['5EE45339', 'A596CE56', '2BD6E687'],
-    goalNotes: [''],
-  },
-  {
-    goalName: 'Ice Cavern HP',
-    trickIds: ['32F255A9', '4674313E', 'B5BDCB36', 'E54BB163'],
-    goalNotes: [''],
-  },
-  {
-    goalName: 'Iron Boots',
-    trickIds: ['32F255A9', '4674313E', 'B5BDCB36', '95759E64'],
-    goalNotes: [''],
-  },
-  {
-    goalName: 'Keaton Mask',
-    trickIds: ['997CEAC9', '7FD1B460'],
-    goalNotes: [
-      "Get Zelda's letter as child",
-      'Give the letter to the Kak Gate guard',
-      'Go to mask shop and get mask',
-    ],
-  },
-  {
-    goalName: 'Lens of Truth',
-    trickIds: ['EAE99180', '7A34B2AD'],
-    goalNotes: [''],
-  },
-  {
-    goalName: 'Light Arrows',
-    trickIds: ['5CC02919', 'A849D697', '3B9CB785'],
-    goalNotes: [
-      'RBA Bugs with broken sword on C right then go to Temple of Time and watch LACs',
-    ],
-  },
-  {
-    goalName: 'Lon Lon Ranch HP',
-    trickIds: ['AE1ED4F9'],
-    goalNotes: [''],
-  },
-  {
-    goalName: 'Longshot',
-    trickIds: ['1E196CEA', 'D7357272', '380D122E'],
-    goalNotes: [''],
-  },
-  {
-    goalName: 'Map & Compass in Bottom of the Well',
-    trickIds: ['622C73E3', 'EAE99180'],
-    goalNotes: [''],
-  },
-  {
-    goalName: 'Map & Compass in Deku Tree',
-    trickIds: ['B669724E'],
-    goalNotes: [''],
-  },
-  {
-    goalName: "Map & Compass in Dodongo's Cavern",
-    trickIds: [],
-    goalNotes: [''],
-  },
-  {
-    goalName: 'Map & Compass in Fire Temple',
-    trickIds: ['8FAE21BA', '6CA81599', '3C261341'],
-    goalNotes: [''],
-  },
-  {
-    goalName: 'Map & Compass in Forest Temple',
-    trickIds: ['9C338AB0', '06950182', '4B71852C', 'D5CEFBA5'],
-    goalNotes: [''],
-  },
-  {
-    goalName: 'Map & Compass in Ice Cavern',
-    trickIds: ['32F255A9', '4674313E', 'B5BDCB36', '8EE322C7'],
-    goalNotes: [''],
-  },
-  {
-    goalName: 'Map & Compass in Jabu-Jabu',
-    trickIds: ['933844BC', '5FCDBB54'],
-    goalNotes: [''],
-  },
-  {
-    goalName: 'Map & Compass in Shadow Temple',
-    trickIds: ['B4E48D8C'],
-    goalNotes: [''],
-  },
-  {
-    goalName: 'Map & Compass in Spirit Temple',
-    trickIds: ['3B9CB785', '8AB11EE5', '375E2213', '42513477', '58D39228'],
-    goalNotes: ['Requires ZL', 'and bow or fire source'],
-  },
-  {
-    goalName: 'Megaton Hammer',
-    trickIds: ['4657BEE7', 'E91970CC', '8FAE21BA', '6CA81599'],
-    goalNotes: [''],
-  },
-  {
-    goalName: 'Milk',
-    trickIds: [],
-    goalNotes: [
-      'Get egg from Malon at Castle',
-      'Wake up Talon',
-      'go to Lon Lon Ranch and win Cucco minigame',
-    ],
-  },
-  {
-    goalName: 'Minuet of Forest',
-    trickIds: ['9C338AB0'],
-    goalNotes: [
-      'Do not RBA broken sword after getting Minuet or you will delete it',
-    ],
-  },
-  {
-    goalName: 'Mirror Shield',
-    trickIds: ['3B9CB785', '8AB11EE5', '375E2213', 'B090EE3A'],
-    goalNotes: [''],
-  },
-  {
-    goalName: "Nayru's Love",
-    trickIds: ['3B9CB785', '8AB11EE5', '375E2213', '5CC02919', 'A849D697'],
-    goalNotes: [
-      'RBA Bugs on B with Poachers Saw on C-Right for ZL',
-      'Get magic at the top of DMT',
-      'then go get NL',
-    ],
-  },
-  {
-    goalName: 'Obtain all 5 Small Keys in Forest Temple',
-    trickIds: ['9C338AB0', '04B44A63'],
-    goalNotes: [''],
-  },
-  {
-    goalName: 'Obtain all 5 Small Keys in Shadow Temple',
-    trickIds: ['B4E48D8C', '075E83B0'],
-    goalNotes: ['Requires ZL for boat'],
-  },
-  {
-    goalName: 'Obtain all 8 Small Keys in Fire Temple',
-    trickIds: ['4657BEE7', '8FAE21BA', '6C288E6E'],
-    goalNotes: [
-      'Can hit rusted twitch under block of time with hammer jumpslash',
-    ],
-  },
-  {
-    goalName: 'Plant bean in Death Mountain Crater',
-    trickIds: ['061F7FDE', '06B33728', '153192BA'],
-    goalNotes: [
-      'Aqua escape from Lost Woods',
-      "buy one magic bean on your way down Zora's River.  Get chus in BOTW",
-      'Climb DMT',
-      'megaflip from top of DMC to the bottom and plan bean.',
-    ],
-  },
-  {
-    goalName: 'Quiver (40)',
-    trickIds: ['5CC02919', 'A849D697'],
-    goalNotes: ['RBA fish with Cojiro on C-right then beat Archery Gallery'],
-  },
-  {
-    goalName: 'Quiver (50)',
-    trickIds: ['5CC02919', 'A849D697', '2ECE35A9'],
-    goalNotes: [
-      'RBA fish with Cojiro on C-right then',
-      'Archery Gallery',
-      'get epona',
-      'get gerudo card and beat horse back archery.  If you steal epona you cannot get caught getting gerudo card. If you do',
-      'she will despawn',
-      'and you will have to go back and re-steal her after getting gerudo card.',
-    ],
-  },
-  {
-    goalName: 'Requiem of Spirit',
-    trickIds: ['3B9CB785', '8AB11EE5', '375E2213'],
-    goalNotes: ['Enter Spirit Temple then Leave Spirit Temple'],
-  },
-  {
-    goalName: "Ruto's Letter",
-    trickIds: ['061F7FDE', '06B33728', '4674313E', 'E92BFD8B'],
-    goalNotes: [
-      'Win divin minigame for scale then take warp into Lake Hylia for letter.',
-    ],
-  },
-  {
-    goalName: "Saria's Song",
-    trickIds: ['997CEAC9', '7FD1B460'],
-    goalNotes: ['Get ZL and then Sarias in Lost Woods'],
-  },
-  {
-    goalName: 'Shadow Temple Boss Key',
-    trickIds: ['202D8126'],
-    goalNotes: ['Requires ZL for boat'],
-  },
-  {
-    goalName: 'Silver Gauntlets',
-    trickIds: ['3B9CB785', '8AB11EE5', '375E2213', 'B090EE3A'],
-    goalNotes: [''],
-  },
-  {
-    goalName: 'Silver Scale',
-    trickIds: ['061F7FDE', '06B33728', '4674313E', 'E92BFD8B'],
-    goalNotes: ['Win diving minigame'],
-  },
-  {
-    goalName: 'Spirit Temple Boss Key',
-    trickIds: ['5CC02919', 'A849D697', '3B9CB785'],
-    goalNotes: [
-      'RBA bugs or fish with light arrows on C-Right for Spirit Boss Key.',
-    ],
-  },
-  {
-    goalName: 'Stone of Agony',
-    trickIds: ['BC67E767'],
-    goalNotes: ['Skulltulas duping allowed', '20 required'],
-  },
-  {
-    goalName: 'Two Fairy Spells',
-    trickIds: [
-      '3B9CB785',
-      '8AB11EE5',
-      '375E2213',
-      '5CC02919',
-      'A849D697',
-      '4674313E',
-      'B5BDCB36',
-    ],
-    goalNotes: ['It is best to get NL and FW as adult and get ZL through RBA.'],
-  },
-  {
-    goalName: 'Win Bombchu Bowling Prize',
-    trickIds: ['2AC9A6C2', 'C664F282'],
-    goalNotes: [
-      'Do not hold z while the bomb chu is exploing overhead',
-      'wait until after it explodes.',
-    ],
-  },
-  {
-    goalName: 'Zora Tunic',
-    trickIds: ['32F255A9', '4674313E', 'B0CA9C91'],
-    goalNotes: [''],
-  },
-  {
-    goalName: 'Clear 4 Silver Rupee Rooms',
-    trickIds: ['FA10A855'],
-    goalNotes: [
-      '- Shadow: turning guillotine room - Shadow: Beamos - Shadow: 2 ReDead room - Ice: turning guillotine - Ice: block room - Spirit: Adult boulder room - Spirit: Child cage room - Spirit: Child sun block room - BotW: basement - GTG: boulder maze - GTG: platforms room - GTG: toilet - Trials: Spirit - Trials: Forest - Trials: Fire - Trials: Light',
-    ],
-  },
-  {
-    goalName: 'Clear 8 Silver Rupee Rooms',
-    trickIds: ['FA10A855'],
-    goalNotes: [
-      '- Shadow: turning guillotine room - Shadow: Beamos - Shadow: 2 ReDead room - Ice: turning guillotine - Ice: block room - Spirit: Adult boulder room - Spirit: Child cage room - Spirit: Child sun block room - BotW: basement - GTG: boulder maze - GTG: platforms room - GTG: toilet - Trials: Spirit - Trials: Forest - Trials: Fire - Trials: Light',
-    ],
-  },
-  {
-    goalName: 'Clear 10 Silver Rupee Rooms',
-    trickIds: ['FA10A855'],
-    goalNotes: [
-      '- Shadow: turning guillotine room - Shadow: Beamos - Shadow: 2 ReDead room - Ice: turning guillotine - Ice: block room - Spirit: Adult boulder room - Spirit: Child cage room - Spirit: Child sun block room - BotW: basement - GTG: boulder maze - GTG: platforms room - GTG: toilet - Trials: Spirit - Trials: Forest - Trials: Fire - Trials: Light',
-    ],
-  },
-  {
-    goalName: 'Open 3 Gold Rupee Chests',
-    trickIds: [
-      '7A34B2AD',
-      'EAE99180',
-      '5CC02919',
-      'A849D697',
-      '89A994DA',
-      '6CA81599',
-      '5EE45339',
-    ],
-    goalNotes: [
-      "BOTW - invisible chest in dead hands rooms | Kak - bombable grotto | Outside Goron City - Song of storms grotto | Goron City - back of rock maze | Fire temple - after skulltula elevator | Gerudo's Training Grounds - Past Silver Gauntlest block  ",
-    ],
-  },
-  {
-    goalName: 'Open 5 Gold Rupee Chests',
-    trickIds: [
-      '7A34B2AD',
-      'EAE99180',
-      '5CC02919',
-      'A849D697',
-      '89A994DA',
-      '6CA81599',
-      '5EE45339',
-    ],
-    goalNotes: [
-      "BOTW - invisible chest in dead hands rooms | Kak - bombable grotto | Outside Goron City - Song of storms grotto | Goron City - back of rock maze | Fire temple - after skulltula elevator | Gerudo's Training Grounds - Past Silver Gauntlest block  ",
-    ],
-  },
-  {
-    goalName: 'Open All 6 Gold Rupee Chests',
-    trickIds: [
-      '7A34B2AD',
-      'EAE99180',
-      '5CC02919',
-      'A849D697',
-      '89A994DA',
-      '6CA81599',
-      '5EE45339',
-    ],
-    goalNotes: [
-      "BOTW - invisible chest in dead hands rooms | Kak - bombable grotto | Outside Goron City - Song of storms grotto | Goron City - back of rock maze | Fire temple - after skulltula elevator | Gerudo's Training Grounds - Past Silver Gauntlest block  ",
-    ],
-  },
-  {
-    goalName: '2 Skulltulas in Lon Lon Ranch',
-    trickIds: [],
-    goalNotes: [''],
-  },
-  {
-    goalName: 'Open 2 Boss Key Doors',
-    trickIds: [
-      '4657BEE7',
-      '7056FCFE',
-      '8A02A837',
-      '70DAAA62',
-      '5EE45339',
-      'A596CE56',
-      '202D8126',
-      'B090EE3A',
-      '537A3654',
-      'FA10A855',
-      'EBBD3E69',
-      '1AD2FFF3',
-    ],
-    goalNotes: ['RBA for water temple BK is usually faster'],
-  },
-  {
-    goalName: 'Open 3 Boss Key Doors',
-    trickIds: [
-      '4657BEE7',
-      '7056FCFE',
-      '8A02A837',
-      '70DAAA62',
-      '5EE45339',
-      'A596CE56',
-      '202D8126',
-      'B090EE3A',
-      '537A3654',
-      'FA10A855',
-      'EBBD3E69',
-      '1AD2FFF3',
-    ],
-    goalNotes: ['RBA for water temple BK is usually faster'],
-  },
-  {
-    goalName: 'Both Rusty Switches in Spirit Temple',
-    trickIds: [
-      '4657BEE7',
-      'E91970CC',
-      '8FAE21BA',
-      '6CA81599',
-      '3B9CB785',
-      '8AB11EE5',
-      '375E2213',
-      '42513477',
-    ],
-    goalNotes: [
-      'Silver Strength+ (basically) Required',
-      'One switch in Main Room',
-      'other through the door that first opens.',
-    ],
-  },
-  {
-    goalName: 'Desert Colossus HP',
-    trickIds: ['3B9CB785', '8AB11EE5', '375E2213', 'B090EE3A', '18268E70'],
-    goalNotes: [''],
-  },
-  {
-    goalName: '20 Different Skulltulas',
-    trickIds: [],
-    goalNotes: [''],
-  },
-  {
-    goalName: '30 Different Skulltulas',
-    trickIds: [],
-    goalNotes: [''],
-  },
-  {
-    goalName: 'Plant 5 Magic Beans',
-    trickIds: [],
-    goalNotes: [
-      'Recommend to plant in Graveyard',
-      'by the bean salesman',
-      'Koriri forest and 2 in lost woods.',
-    ],
-  },
-  {
-    goalName: "All 4 Child Skulltulas in Zora's Domain",
-    trickIds: ['366B4776', '089F7B22'],
-    goalNotes: ['must be night for domain skull'],
-  },
-  {
-    goalName: "All 4 Adult Skulltulas in Zora's Domain",
-    trickIds: ['32F255A9', '4674313E'],
-    goalNotes: ['silver strength+ helps', 'must be night for all skulls'],
-  },
-  {
-    goalName: '7 Different Bottled Contents',
-    trickIds: ['5CC02919', 'A849D697'],
-    goalNotes: [
-      'bugs',
-      'fish',
-      'fairies and blue fire are the only things that can be put in duped bottles',
-    ],
-  },
-  {
-    goalName: 'Open 50 Rupee chest in Wasteland',
-    trickIds: ['3625CB18', 'E9E1A81E', 'D3D2189E', 'D1AE3FD5'],
-    goalNotes: [''],
-  },
-];
