+++ data-access-layer/mock/repositories/mock-trick-repository.ts
@@ -1,22 +1,19 @@
-import {
-  inconsistentMvpGetStartTimeFromUrl,
-  inconsistentMvpGetVideoIdFromUrl,
-} from '../../../components/utils/get_youtube_embed_url';
 import { Trick } from '../../../domain-import-only/Trick';
 import { YoutubeVideo } from '../../../domain-import-only/YoutubeVideo';
 import { ITrickReporistory } from '../../repository-interfaces/trick-repository';
+import { LEGACY_TRICKS, legacy_trick_to_trick } from '../legacy-tricks';
 
 export class MockTrickRepository implements ITrickReporistory {
   async getTrick(trickName: string): Promise<Trick> {
-    const tricksByName = _legacy_tricks()
-      .filter((lt) => lt.trickName == trickName)
-      .map((lt) => {
-        if (lt.trickUrl) {
-          return _legacy_trick_to_trick(lt);
-        } else {
-          return null;
-        }
-      });
+    const tricksByName = LEGACY_TRICKS.filter(
+      (lt) => lt.trickName == trickName
+    ).map((lt) => {
+      if (lt.trickUrl) {
+        return legacy_trick_to_trick(lt);
+      } else {
+        return null;
+      }
+    });
 
     const fallbackTrick: Trick = Trick.create({
       name: 'Fake Trick',
@@ -33,2085 +30,8 @@ export class MockTrickRepository implements ITrickReporistory {
   }
 
   async getAllTricks(): Promise<Trick[]> {
-    return _legacy_tricks().map((lt) => {
-      return _legacy_trick_to_trick(lt);
+    return LEGACY_TRICKS.map((lt) => {
+      return legacy_trick_to_trick(lt);
     });
   }
 }
-
-interface LegacyTrick {
-  id: string;
-  trickName: string;
-  trickUrl?: string;
-  isFundamental: boolean;
-  usuallyRequired: boolean;
-  isMovement: boolean;
-  v1Only: boolean;
-  routing: boolean;
-  practiceSection?: string;
-  notes?: string;
-}
-
-const _legacy_trick_to_trick = (legacyTrick: LegacyTrick) => {
-  let videoId =
-    legacyTrick.trickUrl &&
-    inconsistentMvpGetVideoIdFromUrl(legacyTrick.trickUrl);
-  if (!videoId) {
-    videoId = ''; // just some video
-  }
-
-  let startTime =
-    legacyTrick.trickUrl &&
-    inconsistentMvpGetStartTimeFromUrl(legacyTrick.trickUrl);
-  if (!startTime) {
-    startTime = 0;
-  }
-
-  const trick = Trick.create({
-    id: legacyTrick.id,
-    name: legacyTrick.trickName,
-    video: YoutubeVideo.create({
-      videoId: videoId,
-      startTimeInSeconds: startTime,
-    }),
-    description: legacyTrick.notes,
-    isFundamental: legacyTrick.isFundamental,
-    difficulty: legacyTrick.isFundamental ? 1 : 2,
-  });
-
-  return trick;
-};
-
-/** PREVIOUS SOURCE DATA */
-const _legacy_tricks = () => {
-  return [
-    {
-      id: 'CF6E0DAD',
-      trickName: 'Pokey Escape',
-      trickUrl: null,
-      isFundamental: true,
-      usuallyRequired: true,
-      isMovement: false,
-      v1Only: false,
-      routing: false,
-      practiceSection: 'Escape',
-      notes: null,
-    },
-    {
-      id: '470DFD19',
-      trickName: 'Aqua Escape',
-      trickUrl: null,
-      isFundamental: true,
-      usuallyRequired: true,
-      isMovement: false,
-      v1Only: false,
-      routing: false,
-      practiceSection: 'Escape',
-      notes: null,
-    },
-    {
-      id: '81FAD7D4',
-      trickName: 'Walking While Talking Escape',
-      trickUrl: null,
-      isFundamental: true,
-      usuallyRequired: true,
-      isMovement: false,
-      v1Only: true,
-      routing: false,
-      practiceSection: 'Escape',
-      notes: null,
-    },
-    {
-      id: 'EAE99180',
-      trickName: 'Navi Dive',
-      trickUrl: null,
-      isFundamental: true,
-      usuallyRequired: true,
-      isMovement: true,
-      v1Only: false,
-      routing: false,
-      practiceSection: 'Escape',
-      notes: null,
-    },
-    {
-      id: 'C15A5A39',
-      trickName: 'Castle Before Night Unloaded Chains',
-      trickUrl: null,
-      isFundamental: true,
-      usuallyRequired: true,
-      isMovement: false,
-      v1Only: false,
-      routing: false,
-      practiceSection: 'Market/Castle',
-      notes: null,
-    },
-    {
-      id: 'B63CCD62',
-      trickName: 'Baazar at Night',
-      trickUrl: null,
-      isFundamental: true,
-      usuallyRequired: true,
-      isMovement: false,
-      v1Only: false,
-      routing: false,
-      practiceSection: 'Market/Castle',
-      notes: null,
-    },
-    {
-      id: '997CEAC9',
-      trickName: 'Seam Walk + Chu Jump To Guards',
-      trickUrl: null,
-      isFundamental: true,
-      usuallyRequired: false,
-      isMovement: false,
-      v1Only: false,
-      routing: false,
-      practiceSection: 'Market/Castle',
-      notes: null,
-    },
-    {
-      id: '7FD1B460',
-      trickName: 'ZL CS Skip',
-      trickUrl: null,
-      isFundamental: true,
-      usuallyRequired: false,
-      isMovement: false,
-      v1Only: false,
-      routing: false,
-      practiceSection: 'Market/Castle',
-      notes: null,
-    },
-    {
-      id: '65A2A17C',
-      trickName: 'Child DoT Skip',
-      trickUrl: null,
-      isFundamental: true,
-      usuallyRequired: true,
-      isMovement: false,
-      v1Only: false,
-      routing: false,
-      practiceSection: 'Market/Castle',
-      notes: null,
-    },
-    {
-      id: '048B036A',
-      trickName: 'Song of Storms Grotto Skulltchula',
-      trickUrl: 'https://www.youtube.com/watch?v=Z8X27ZrPBeo',
-      isFundamental: false,
-      usuallyRequired: false,
-      isMovement: false,
-      v1Only: false,
-      routing: false,
-      practiceSection: 'Market/Castle',
-      notes: null,
-    },
-    {
-      id: '537A3654',
-      trickName: 'After Hit Invinsibility',
-      trickUrl: null,
-      isFundamental: true,
-      usuallyRequired: false,
-      isMovement: true,
-      v1Only: false,
-      routing: false,
-      practiceSection: 'Pure Movement',
-      notes: null,
-    },
-    {
-      id: 'A47278E1',
-      trickName: 'ESS Roll',
-      trickUrl: null,
-      isFundamental: true,
-      usuallyRequired: false,
-      isMovement: true,
-      v1Only: false,
-      routing: false,
-      practiceSection: 'Pure Movement',
-      notes: null,
-    },
-    {
-      id: '310D8018',
-      trickName: 'Weirdshot',
-      trickUrl:
-        'https://youtu.be/T57Hfh3bWr8?list=TLPQMjAwNjIwMjCT5KarRPwM5g&t=72',
-      isFundamental: true,
-      usuallyRequired: false,
-      isMovement: true,
-      v1Only: false,
-      routing: false,
-      practiceSection: 'Pure Movement',
-      notes: null,
-    },
-    {
-      id: 'CCF4A6DB',
-      trickName: 'Ground Jump (Bombs)',
-      trickUrl: null,
-      isFundamental: true,
-      usuallyRequired: true,
-      isMovement: true,
-      v1Only: false,
-      routing: false,
-      practiceSection: 'Pure Movement',
-      notes: null,
-    },
-    {
-      id: 'BFE8E49C',
-      trickName: 'Mega sidehop (Chu and Bomb Strats)',
-      trickUrl: null,
-      isFundamental: true,
-      usuallyRequired: true,
-      isMovement: true,
-      v1Only: false,
-      routing: false,
-      practiceSection: 'Pure Movement',
-      notes: null,
-    },
-    {
-      id: 'D0D4BC13',
-      trickName: 'Superslide (Bombs)',
-      trickUrl: null,
-      isFundamental: true,
-      usuallyRequired: true,
-      isMovement: true,
-      v1Only: false,
-      routing: false,
-      practiceSection: 'Pure Movement',
-      notes: null,
-    },
-    {
-      id: 'AED0754A',
-      trickName: 'WESS',
-      trickUrl: null,
-      isFundamental: true,
-      usuallyRequired: true,
-      isMovement: true,
-      v1Only: false,
-      routing: false,
-      practiceSection: 'Pure Movement',
-      notes: null,
-    },
-    {
-      id: '8EE322C7',
-      trickName: 'Ledge Clip',
-      trickUrl: null,
-      isFundamental: true,
-      usuallyRequired: false,
-      isMovement: true,
-      v1Only: false,
-      routing: false,
-      practiceSection: 'Pure Movement',
-      notes: null,
-    },
-    {
-      id: '6F7BC03E',
-      trickName: 'Sideflip Hovering',
-      trickUrl: null,
-      isFundamental: true,
-      usuallyRequired: false,
-      isMovement: true,
-      v1Only: false,
-      routing: false,
-      practiceSection: 'Pure Movement',
-      notes: null,
-    },
-    {
-      id: 'B0C59F81',
-      trickName: 'Child hover to skull',
-      trickUrl: 'https://www.youtube.com/watch?v=Rv7TNiiq34U',
-      isFundamental: false,
-      usuallyRequired: false,
-      isMovement: false,
-      v1Only: false,
-      routing: false,
-      practiceSection: 'Valley-Wasteland',
-      notes: null,
-    },
-    {
-      id: '9680FAC3',
-      trickName: '37th Heartpiece',
-      trickUrl: 'https://youtu.be/QYbEfuYUcmQ',
-      isFundamental: false,
-      usuallyRequired: false,
-      isMovement: false,
-      v1Only: false,
-      routing: false,
-      practiceSection: 'Valley-Wasteland',
-      notes: null,
-    },
-    {
-      id: '3B9CB785',
-      trickName: 'Cross Gerudo Bridge',
-      trickUrl: 'https://www.youtube.com/watch?v=0i7sXA1OtYk',
-      isFundamental: true,
-      usuallyRequired: true,
-      isMovement: false,
-      v1Only: false,
-      routing: false,
-      practiceSection: 'Valley-Wasteland',
-      notes: null,
-    },
-    {
-      id: '8AB11EE5',
-      trickName: 'Gerduo Gate Skip',
-      trickUrl: 'https://www.youtube.com/watch?v=jWpI0jIk9oI&t=189s',
-      isFundamental: true,
-      usuallyRequired: true,
-      isMovement: false,
-      v1Only: false,
-      routing: false,
-      practiceSection: 'Valley-Wasteland',
-      notes: null,
-    },
-    {
-      id: '3625CB18',
-      trickName: 'Child Bridge Megaflip',
-      trickUrl: 'https://youtu.be/tMl9vYDnjjQ',
-      isFundamental: false,
-      usuallyRequired: false,
-      isMovement: false,
-      v1Only: false,
-      routing: false,
-      practiceSection: 'Valley-Wasteland',
-      notes:
-        'release the stick after megaflipping to avoid climbing the gate (less likely to get caught)',
-    },
-    {
-      id: 'E9E1A81E',
-      trickName: 'Child Gate Skip',
-      trickUrl: 'https://youtu.be/qh0zt_S4ykQ',
-      isFundamental: false,
-      usuallyRequired: false,
-      isMovement: false,
-      v1Only: false,
-      routing: false,
-      practiceSection: 'Valley-Wasteland',
-      notes: null,
-    },
-    {
-      id: '375E2213',
-      trickName: 'Wasteland No Lens',
-      trickUrl: 'https://www.youtube.com/watch?v=h5_EiPTjdEk',
-      isFundamental: true,
-      usuallyRequired: true,
-      isMovement: false,
-      v1Only: false,
-      routing: false,
-      practiceSection: 'Valley-Wasteland',
-      notes: null,
-    },
-    {
-      id: '6F1BB8D3',
-      trickName: 'Weirdclip to GS Near Gerudo Valley',
-      trickUrl: 'https://www.youtube.com/watch?v=8jR81paCotM',
-      isFundamental: false,
-      usuallyRequired: false,
-      isMovement: false,
-      v1Only: false,
-      routing: false,
-      practiceSection: 'Valley-Wasteland',
-      notes: null,
-    },
-    {
-      id: 'D3D2189E',
-      trickName: 'River of Sand No Hoverboots',
-      trickUrl: 'https://youtu.be/x1rIszTnER8',
-      isFundamental: false,
-      usuallyRequired: false,
-      isMovement: false,
-      v1Only: false,
-      routing: false,
-      practiceSection: 'Valley-Wasteland',
-      notes: null,
-    },
-    {
-      id: 'D1AE3FD5',
-      trickName: 'Wasteland 50 rupee chest',
-      trickUrl: 'https://www.youtube.com/watch?v=ttWVO4aGBH4',
-      isFundamental: false,
-      usuallyRequired: false,
-      isMovement: false,
-      v1Only: false,
-      routing: false,
-      practiceSection: 'Valley-Wasteland',
-      notes: null,
-    },
-    {
-      id: 'BA12A4D9',
-      trickName: 'All 5 DC Skulls Route',
-      trickUrl:
-        'https://www.youtube.com/watch?v=dpZUBS1jFy0&list=PLis_W6MEPku3PAobg6W1zP0-N8j7e-l08&index=8&t=0s',
-      isFundamental: false,
-      usuallyRequired: false,
-      isMovement: false,
-      v1Only: false,
-      routing: false,
-      practiceSection: "Dodongo's Cavern",
-      notes: null,
-    },
-    {
-      id: 'B69C4380',
-      trickName: 'Doomjump',
-      trickUrl: 'https://www.youtube.com/watch?v=WZNVlNNYmx8&t=58s',
-      isFundamental: false,
-      usuallyRequired: false,
-      isMovement: false,
-      v1Only: false,
-      routing: false,
-      practiceSection: "Dodongo's Cavern",
-      notes: null,
-    },
-    {
-      id: '6C2A9F69',
-      trickName: 'Dodongo Door Block Skip',
-      trickUrl: 'https://www.youtube.com/watch?v=0kMaQQ_-5Ds&t=19s',
-      isFundamental: false,
-      usuallyRequired: false,
-      isMovement: false,
-      v1Only: false,
-      routing: false,
-      practiceSection: "Dodongo's Cavern",
-      notes: null,
-    },
-    {
-      id: 'C664F282',
-      trickName: 'DC WW to GTG as Child',
-      trickUrl:
-        'https://www.youtube.com/watch?v=0ZZ6fRs5gnI&list=PL1PISDP98fygAnj3yMpFfP4IWrPs9NpqJ&index=20&t=0s',
-      isFundamental: false,
-      usuallyRequired: false,
-      isMovement: false,
-      v1Only: false,
-      routing: false,
-      practiceSection: "Dodongo's Cavern",
-      notes: null,
-    },
-    {
-      id: 'E91970CC',
-      trickName: 'Fastest Pure Hammer Route',
-      trickUrl: 'https://youtu.be/EurMMhKXklA?t=4612',
-      isFundamental: false,
-      usuallyRequired: false,
-      isMovement: false,
-      v1Only: false,
-      routing: false,
-      practiceSection: 'Fire Temple',
-      notes: null,
-    },
-    {
-      id: '6CA81599',
-      trickName: 'Fire Temple Push Block Skip',
-      trickUrl:
-        'https://www.youtube.com/watch?v=V8e5C17Olfg&list=PLis_W6MEPku3muvelCmPhgizb2FpC1V9z&index=4&t=0s',
-      isFundamental: false,
-      usuallyRequired: false,
-      isMovement: false,
-      v1Only: false,
-      routing: false,
-      practiceSection: 'Fire Temple',
-      notes: null,
-    },
-    {
-      id: '6C288E6E',
-      trickName: 'All 5 Fire Skulls Route',
-      trickUrl:
-        'https://www.youtube.com/watch?v=KS8o7OILi44&list=PLis_W6MEPku3muvelCmPhgizb2FpC1V9z&index=2&t=0s',
-      isFundamental: false,
-      usuallyRequired: false,
-      isMovement: false,
-      v1Only: false,
-      routing: false,
-      practiceSection: 'Fire Temple',
-      notes: null,
-    },
-    {
-      id: '8FAE21BA',
-      trickName: 'Darunia CS Skip',
-      trickUrl: 'https://www.youtube.com/watch?v=hBifIqV3hyc',
-      isFundamental: false,
-      usuallyRequired: false,
-      isMovement: false,
-      v1Only: false,
-      routing: false,
-      practiceSection: 'Fire Temple',
-      notes: null,
-    },
-    {
-      id: '7056FCFE',
-      trickName: 'Fire Weird Shot to BK',
-      trickUrl:
-        'https://www.youtube.com/watch?v=Ms_sfKav1kA&list=PLis_W6MEPku3muvelCmPhgizb2FpC1V9z&index=5&t=0s',
-      isFundamental: false,
-      usuallyRequired: false,
-      isMovement: false,
-      v1Only: false,
-      routing: false,
-      practiceSection: 'Fire Temple',
-      notes: null,
-    },
-    {
-      id: '11DF7000',
-      trickName: 'Fire SoT Block Skip For Skull',
-      trickUrl:
-        'https://www.youtube.com/watch?v=dkAx8gqevpE&list=PLis_W6MEPku3muvelCmPhgizb2FpC1V9z&index=6&t=0s',
-      isFundamental: false,
-      usuallyRequired: false,
-      isMovement: false,
-      v1Only: false,
-      routing: false,
-      practiceSection: 'Fire Temple',
-      notes: null,
-    },
-    {
-      id: '89A994DA',
-      trickName: 'Fire Hover to Elevator for 2 skulls',
-      trickUrl:
-        'https://www.youtube.com/watch?v=f9ZIyYMM6zI&list=PLis_W6MEPku3muvelCmPhgizb2FpC1V9z&index=7&t=0s',
-      isFundamental: false,
-      usuallyRequired: false,
-      isMovement: false,
-      v1Only: false,
-      routing: false,
-      practiceSection: 'Fire Temple',
-      notes: null,
-    },
-    {
-      id: '3C261341',
-      trickName: 'Map and Compass in Fire',
-      trickUrl:
-        'https://youtu.be/PfGFYaQQkB4?list=PLis_W6MEPku3muvelCmPhgizb2FpC1V9z&t=50',
-      isFundamental: false,
-      usuallyRequired: false,
-      isMovement: false,
-      v1Only: false,
-      routing: false,
-      practiceSection: 'Fire Temple',
-      notes: null,
-    },
-    {
-      id: 'E7A8AC29',
-      trickName: 'Volvagia 1 Cycle',
-      trickUrl:
-        'https://www.youtube.com/watch?v=dcl2P53vL2s&list=PLis_W6MEPku3muvelCmPhgizb2FpC1V9z&index=9&t=180s',
-      isFundamental: false,
-      usuallyRequired: false,
-      isMovement: false,
-      v1Only: false,
-      routing: false,
-      practiceSection: 'Fire Temple',
-      notes: null,
-    },
-    {
-      id: '815D2CE3',
-      trickName: 'Fire WW to Forest',
-      trickUrl:
-        'https://www.youtube.com/watch?v=A074PAKb_ac&list=PLis_W6MEPku3muvelCmPhgizb2FpC1V9z&index=10&t=42s',
-      isFundamental: false,
-      usuallyRequired: false,
-      isMovement: false,
-      v1Only: false,
-      routing: false,
-      practiceSection: 'Fire Temple',
-      notes: null,
-    },
-    {
-      id: 'C036348D',
-      trickName: 'Fire Megasidehop BK Skip ',
-      trickUrl:
-        'https://www.youtube.com/watch?v=YIHOw0i-JHI&list=PLis_W6MEPku3muvelCmPhgizb2FpC1V9z&index=11&t=0s',
-      isFundamental: false,
-      usuallyRequired: false,
-      isMovement: false,
-      v1Only: false,
-      routing: false,
-      practiceSection: 'Fire Temple',
-      notes: null,
-    },
-    {
-      id: '76C797C9',
-      trickName: 'Fire Ledge Clip BK Skip ',
-      trickUrl:
-        'https://www.youtube.com/watch?v=s-gbowCh584&list=PLis_W6MEPku3muvelCmPhgizb2FpC1V9z&index=12&t=0s',
-      isFundamental: false,
-      usuallyRequired: false,
-      isMovement: false,
-      v1Only: false,
-      routing: false,
-      practiceSection: 'Fire Temple',
-      notes: null,
-    },
-    {
-      id: '8051E7C4',
-      trickName: 'Fire HESS BK Skip',
-      trickUrl: 'https://youtu.be/K3srLjTMqm4',
-      isFundamental: false,
-      usuallyRequired: false,
-      isMovement: false,
-      v1Only: false,
-      routing: false,
-      practiceSection: 'Fire Temple',
-      notes: null,
-    },
-    {
-      id: '04B44A63',
-      trickName: 'Fairy Bow Route',
-      trickUrl: 'https://youtu.be/PBo7iBtta6c?t=3113',
-      isFundamental: false,
-      usuallyRequired: false,
-      isMovement: false,
-      v1Only: false,
-      routing: false,
-      practiceSection: 'Forest Temple',
-      notes: null,
-    },
-    {
-      id: 'ED22ECBC',
-      trickName: 'Forest Temple BK Skip',
-      trickUrl: 'https://www.youtube.com/watch?v=5DfkXaHgeG4',
-      isFundamental: false,
-      usuallyRequired: false,
-      isMovement: false,
-      v1Only: false,
-      routing: false,
-      practiceSection: 'Forest Temple',
-      notes: null,
-    },
-    {
-      id: '4B71852C',
-      trickName: 'Forest Temple SoT Block Clip',
-      trickUrl: 'https://youtu.be/jI2ZzvkmmBw?t=3',
-      isFundamental: false,
-      usuallyRequired: false,
-      isMovement: false,
-      v1Only: false,
-      routing: false,
-      practiceSection: 'Forest Temple',
-      notes: null,
-    },
-    {
-      id: '06950182',
-      trickName: 'Forest Megajump',
-      trickUrl: 'https://www.youtube.com/watch?v=yWNIkOXYWO4',
-      isFundamental: false,
-      usuallyRequired: false,
-      isMovement: false,
-      v1Only: false,
-      routing: false,
-      practiceSection: 'Forest Temple',
-      notes: null,
-    },
-    {
-      id: '03615B32',
-      trickName: 'Forest Temple Poe CS Skip',
-      trickUrl: 'https://www.youtube.com/watch?v=x0jM5jU-J-M&feature=youtu.be',
-      isFundamental: false,
-      usuallyRequired: false,
-      isMovement: false,
-      v1Only: false,
-      routing: false,
-      practiceSection: 'Forest Temple',
-      notes: null,
-    },
-    {
-      id: 'D5CEFBA5',
-      trickName: 'Forest Block Push Puzzle Skip',
-      trickUrl: 'https://www.youtube.com/watch?v=dIcQMuPA4mc',
-      isFundamental: false,
-      usuallyRequired: false,
-      isMovement: false,
-      v1Only: false,
-      routing: false,
-      practiceSection: 'Forest Temple',
-      notes: null,
-    },
-    {
-      id: '9619C13C',
-      trickName: 'All 5 Forest Skulls Route',
-      trickUrl:
-        'https://www.youtube.com/watch?v=fztkhqz-C5k&list=PLis_W6MEPku3PAobg6W1zP0-N8j7e-l08&index=2&t=0s',
-      isFundamental: false,
-      usuallyRequired: false,
-      isMovement: false,
-      v1Only: false,
-      routing: false,
-      practiceSection: 'Forest Temple',
-      notes: null,
-    },
-    {
-      id: 'E981EB6E',
-      trickName: 'Forest Impossible Jump',
-      trickUrl: 'https://www.youtube.com/watch?v=neH1M8CqL2Q',
-      isFundamental: false,
-      usuallyRequired: false,
-      isMovement: false,
-      v1Only: false,
-      routing: false,
-      practiceSection: 'Forest Temple',
-      notes: null,
-    },
-    {
-      id: '0A1DD7ED',
-      trickName: 'Forest Medallion CS Skip (Bomb)',
-      trickUrl: 'https://www.youtube.com/watch?v=f74txDcMiUA&feature=youtu.be',
-      isFundamental: false,
-      usuallyRequired: false,
-      isMovement: false,
-      v1Only: false,
-      routing: false,
-      practiceSection: 'Forest Temple',
-      notes: null,
-    },
-    {
-      id: 'A9F572CC',
-      trickName: 'Phantom Ganon Strategy',
-      trickUrl: 'https://youtu.be/1lqAeNwt5z8',
-      isFundamental: false,
-      usuallyRequired: false,
-      isMovement: false,
-      v1Only: false,
-      routing: false,
-      practiceSection: 'Forest Temple',
-      notes: null,
-    },
-    {
-      id: 'ACBB4ECB',
-      trickName: 'Hover Boost To Amy',
-      trickUrl: 'https://youtu.be/p4TgxwZcgyg',
-      isFundamental: false,
-      usuallyRequired: false,
-      isMovement: false,
-      v1Only: false,
-      routing: false,
-      practiceSection: 'Forest Temple',
-      notes: null,
-    },
-    {
-      id: '9C338AB0',
-      trickName: 'Minuet CS Skip',
-      trickUrl: 'https://www.youtube.com/watch?v=Z4p-ZQX2XGA',
-      isFundamental: true,
-      usuallyRequired: true,
-      isMovement: false,
-      v1Only: false,
-      routing: false,
-      practiceSection: 'Forest Temple',
-      notes: null,
-    },
-    {
-      id: '8A02A837',
-      trickName: 'Forest Boss Door Route',
-      trickUrl:
-        'https://www.youtube.com/watch?time_continue=1&v=qteeWesUmYE&feature=emb_logo',
-      isFundamental: false,
-      usuallyRequired: false,
-      isMovement: false,
-      v1Only: false,
-      routing: false,
-      practiceSection: 'Forest Temple',
-      notes: null,
-    },
-    {
-      id: '1E196CEA',
-      trickName: 'Clip into Water Temple from Serenade',
-      trickUrl: 'https://www.youtube.com/watch?v=AqdCphIIP-I',
-      isFundamental: false,
-      usuallyRequired: false,
-      isMovement: false,
-      v1Only: false,
-      routing: false,
-      practiceSection: 'Water Temple',
-      notes: null,
-    },
-    {
-      id: 'D7357272',
-      trickName: 'Clip into Water From Lab',
-      trickUrl: 'https://www.youtube.com/watch?v=nkz1suNoI1s',
-      isFundamental: false,
-      usuallyRequired: false,
-      isMovement: false,
-      v1Only: false,
-      routing: false,
-      practiceSection: 'Water Temple',
-      notes: null,
-    },
-    {
-      id: '85D1C987',
-      trickName: '3 Skulls in Water Route',
-      trickUrl:
-        'https://www.youtube.com/watch?v=jLWPQ__MO4A&list=PLis_W6MEPku0v7MhoxE09yGBnsDuh5bLL&index=13&t=22s',
-      isFundamental: false,
-      usuallyRequired: false,
-      isMovement: false,
-      v1Only: false,
-      routing: false,
-      practiceSection: 'Water Temple',
-      notes: null,
-    },
-    {
-      id: '6DD78777',
-      trickName: 'Water BK Skip',
-      trickUrl:
-        'https://www.youtube.com/watch?v=g3wJXxRcUJ0&list=PLis_W6MEPku0v7MhoxE09yGBnsDuh5bLL&index=6&t=0s',
-      isFundamental: false,
-      usuallyRequired: false,
-      isMovement: false,
-      v1Only: false,
-      routing: false,
-      practiceSection: 'Water Temple',
-      notes: null,
-    },
-    {
-      id: '70DAAA62',
-      trickName: 'Water Superslide To Boss Room',
-      trickUrl:
-        'https://www.youtube.com/watch?v=Q-DSL3vW2Xk&list=PLis_W6MEPku0v7MhoxE09yGBnsDuh5bLL&index=5&t=0s',
-      isFundamental: false,
-      usuallyRequired: false,
-      isMovement: false,
-      v1Only: false,
-      routing: false,
-      practiceSection: 'Water Temple',
-      notes: null,
-    },
-    {
-      id: '771D2D81',
-      trickName: 'Water Compass Route',
-      trickUrl:
-        'https://www.youtube.com/watch?v=JSAi2WNnisw&list=PLis_W6MEPku0v7MhoxE09yGBnsDuh5bLL&index=11&t=0s',
-      isFundamental: false,
-      usuallyRequired: false,
-      isMovement: false,
-      v1Only: false,
-      routing: false,
-      practiceSection: 'Water Temple',
-      notes: null,
-    },
-    {
-      id: 'C3FBA8DC',
-      trickName: 'Water Medallion CS Skip',
-      trickUrl:
-        'https://www.youtube.com/watch?v=q5TPOHWw5TM&list=PLis_W6MEPku0v7MhoxE09yGBnsDuh5bLL&index=10&t=20s',
-      isFundamental: false,
-      usuallyRequired: false,
-      isMovement: false,
-      v1Only: false,
-      routing: false,
-      practiceSection: 'Water Temple',
-      notes: null,
-    },
-    {
-      id: '3AAE8516',
-      trickName: 'SoT Block Skip',
-      trickUrl:
-        'https://www.youtube.com/watch?v=wkQyTYoFsL4&list=PLis_W6MEPku0v7MhoxE09yGBnsDuh5bLL&index=4&t=0s',
-      isFundamental: false,
-      usuallyRequired: false,
-      isMovement: false,
-      v1Only: false,
-      routing: false,
-      practiceSection: 'Water Temple',
-      notes: null,
-    },
-    {
-      id: '7CB759AA',
-      trickName: 'All 5 Skulls (Iron Boots)',
-      trickUrl:
-        'https://www.youtube.com/watch?v=LRGKm5CwGJU&list=PLis_W6MEPku3PAobg6W1zP0-N8j7e-l08&index=4&t=0s',
-      isFundamental: false,
-      usuallyRequired: false,
-      isMovement: false,
-      v1Only: false,
-      routing: false,
-      practiceSection: 'Water Temple',
-      notes: null,
-    },
-    {
-      id: '120D7D11',
-      trickName: 'All 5 Skulls (ZL+Bow)',
-      trickUrl:
-        'https://www.youtube.com/watch?v=ShEPxI7qkMU&list=PLis_W6MEPku3PAobg6W1zP0-N8j7e-l08&index=5&t=0s',
-      isFundamental: false,
-      usuallyRequired: false,
-      isMovement: false,
-      v1Only: false,
-      routing: false,
-      practiceSection: 'Water Temple',
-      notes: null,
-    },
-    {
-      id: '380D122E',
-      trickName: 'Dark Link Strategy',
-      trickUrl: 'https://www.youtube.com/watch?v=X1Aou84VmXc',
-      isFundamental: false,
-      usuallyRequired: false,
-      isMovement: false,
-      v1Only: false,
-      routing: false,
-      practiceSection: 'Water Temple',
-      notes: null,
-    },
-    {
-      id: 'DD4D7F1C',
-      trickName: 'Morpha Quick Kill Strategy',
-      trickUrl: 'https://youtu.be/u9Hs23Xwo-w?t=376',
-      isFundamental: false,
-      usuallyRequired: false,
-      isMovement: false,
-      v1Only: false,
-      routing: false,
-      practiceSection: 'Water Temple',
-      notes: null,
-    },
-    {
-      id: 'D80E6EDA',
-      trickName: 'Water Map From Compass Room',
-      trickUrl: 'https://www.youtube.com/watch?v=6-EIF3GOJJk',
-      isFundamental: false,
-      usuallyRequired: false,
-      isMovement: false,
-      v1Only: false,
-      routing: false,
-      practiceSection: 'Water Temple',
-      notes: null,
-    },
-    {
-      id: '4E0619E0',
-      trickName: 'One Water Key (No ZL, No Iron Boots)',
-      trickUrl: 'https://www.youtube.com/watch?v=GjUgW0W4-oY',
-      isFundamental: false,
-      usuallyRequired: false,
-      isMovement: false,
-      v1Only: false,
-      routing: false,
-      practiceSection: 'Water Temple',
-      notes: null,
-    },
-    {
-      id: 'C51AA108',
-      trickName: 'Sidehop to bombable wall',
-      trickUrl: 'https://youtu.be/dszbnOhodwU',
-      isFundamental: false,
-      usuallyRequired: false,
-      isMovement: false,
-      v1Only: false,
-      routing: false,
-      practiceSection: 'Water Temple',
-      notes: null,
-    },
-    {
-      id: 'B4E48D8C',
-      trickName: 'Shadow Temple Early (Hookshot Jump)',
-      trickUrl: 'https://www.youtube.com/watch?v=HTU0fKl-6uQ&t=4s',
-      isFundamental: true,
-      usuallyRequired: true,
-      isMovement: false,
-      v1Only: false,
-      routing: false,
-      practiceSection: 'Shadow Temple',
-      notes: null,
-    },
-    {
-      id: 'B6828BD7',
-      trickName: 'All 5 Shadow Temple Skulls Route',
-      trickUrl:
-        'https://www.youtube.com/watch?v=_S805lyNKcg&list=PLis_W6MEPku3PAobg6W1zP0-N8j7e-l08&index=6&t=0s',
-      isFundamental: false,
-      usuallyRequired: false,
-      isMovement: false,
-      v1Only: false,
-      routing: false,
-      practiceSection: 'Shadow Temple',
-      notes: null,
-    },
-    {
-      id: '075E83B0',
-      trickName: 'Shadow Fence Clip',
-      trickUrl: 'https://www.youtube.com/watch?v=udCyLj2brsc&feature=youtu.be',
-      isFundamental: false,
-      usuallyRequired: false,
-      isMovement: false,
-      v1Only: false,
-      routing: false,
-      practiceSection: 'Shadow Temple',
-      notes: null,
-    },
-    {
-      id: '00D651EC',
-      trickName: 'Shadow Skull in Boat Room',
-      trickUrl: 'https://www.youtube.com/watch?v=oX-dmoggudk',
-      isFundamental: false,
-      usuallyRequired: false,
-      isMovement: false,
-      v1Only: false,
-      routing: false,
-      practiceSection: 'Shadow Temple',
-      notes: null,
-    },
-    {
-      id: '202D8126',
-      trickName: 'Shadow BK No Dins',
-      trickUrl: 'https://youtu.be/lI1nlIbzoKw',
-      isFundamental: false,
-      usuallyRequired: false,
-      isMovement: false,
-      v1Only: false,
-      routing: false,
-      practiceSection: 'Shadow Temple',
-      notes: null,
-    },
-    {
-      id: '87D42917',
-      trickName: 'Boatkey Skip',
-      trickUrl:
-        'https://www.youtube.com/watch?v=sRf8wiaD02o&list=PL1n_Y5KrDdmUu3R2WUXPEmKjsg1XEr84u&index=9&t=0s',
-      isFundamental: false,
-      usuallyRequired: false,
-      isMovement: false,
-      v1Only: false,
-      routing: false,
-      practiceSection: 'Shadow Temple',
-      notes: null,
-    },
-    {
-      id: '5BAD1B80',
-      trickName: 'Shadow Spike Crusher Room',
-      trickUrl: 'https://www.youtube.com/watch?v=gYUXB834PvI',
-      isFundamental: false,
-      usuallyRequired: false,
-      isMovement: false,
-      v1Only: false,
-      routing: false,
-      practiceSection: 'Shadow Temple',
-      notes: null,
-    },
-    {
-      id: 'DB6FF011',
-      trickName: '(Non 1.0) Bongo Strategy',
-      trickUrl: 'https://youtu.be/AgBByp1nCl4?t=17767',
-      isFundamental: false,
-      usuallyRequired: false,
-      isMovement: false,
-      v1Only: false,
-      routing: false,
-      practiceSection: 'Shadow Temple',
-      notes: null,
-    },
-    {
-      id: '932822FE',
-      trickName: 'Shadow Bow Skip to Cross Void',
-      trickUrl: 'https://www.youtube.com/watch?v=bG9oyzWRX40',
-      isFundamental: false,
-      usuallyRequired: false,
-      isMovement: false,
-      v1Only: false,
-      routing: false,
-      practiceSection: 'Shadow Temple',
-      notes: null,
-    },
-    {
-      id: '1F23C2BE',
-      trickName: 'Shadow Small Key Route',
-      trickUrl: 'https://www.youtube.com/watch?v=5QpJ1YV3gDo',
-      isFundamental: false,
-      usuallyRequired: false,
-      isMovement: false,
-      v1Only: false,
-      routing: false,
-      practiceSection: 'Shadow Temple',
-      notes: null,
-    },
-    {
-      id: '8E98B6FC',
-      trickName: 'Shadow Medallion CS Skip',
-      trickUrl:
-        'https://www.youtube.com/watch?v=WVZFy0xdMVI&list=PL1n_Y5KrDdmUu3R2WUXPEmKjsg1XEr84u&index=17&t=0s',
-      isFundamental: false,
-      usuallyRequired: false,
-      isMovement: false,
-      v1Only: false,
-      routing: false,
-      practiceSection: 'Shadow Temple',
-      notes: null,
-    },
-    {
-      id: 'A0C2A8B5',
-      trickName: '(1.0 Only) Bongo Quick Kill',
-      trickUrl: 'https://youtu.be/AgBByp1nCl4?t=7823',
-      isFundamental: false,
-      usuallyRequired: false,
-      isMovement: false,
-      v1Only: true,
-      routing: false,
-      practiceSection: 'Shadow Temple',
-      notes: null,
-    },
-    {
-      id: '614800B4',
-      trickName: 'All 5 Spirit Skulls Route',
-      trickUrl:
-        'https://www.youtube.com/watch?v=eeT4BRTQN8Y&list=PLis_W6MEPku3PAobg6W1zP0-N8j7e-l08&index=7&t=110s',
-      isFundamental: false,
-      usuallyRequired: false,
-      isMovement: false,
-      v1Only: false,
-      routing: false,
-      practiceSection: 'Spirit Temple',
-      notes: null,
-    },
-    {
-      id: 'B090EE3A',
-      trickName: 'Spirit Hover',
-      trickUrl:
-        'https://www.youtube.com/watch?v=L_i3Q_1FDqM&list=PLis_W6MEPku2We7VhrYe9i-2iEzbg6JJh&index=22&t=0s',
-      isFundamental: false,
-      usuallyRequired: false,
-      isMovement: false,
-      v1Only: false,
-      routing: false,
-      practiceSection: 'Spirit Temple',
-      notes: null,
-    },
-    {
-      id: '18268E70',
-      trickName: 'Slide to Colossus HP',
-      trickUrl: 'https://www.youtube.com/watch?v=kkzgClM3Q5w',
-      isFundamental: false,
-      usuallyRequired: false,
-      isMovement: false,
-      v1Only: false,
-      routing: false,
-      practiceSection: 'Spirit Temple',
-      notes: null,
-    },
-    {
-      id: 'E33B98C0',
-      trickName: 'Statue Climb BK Skip',
-      trickUrl: 'https://www.youtube.com/watch?v=CF_PRShnFGE',
-      isFundamental: false,
-      usuallyRequired: false,
-      isMovement: false,
-      v1Only: false,
-      routing: false,
-      practiceSection: 'Spirit Temple',
-      notes: null,
-    },
-    {
-      id: 'FB73AEF8',
-      trickName: 'Twinrova Fight',
-      trickUrl: 'https://youtu.be/AgBByp1nCl4?t=10855',
-      isFundamental: false,
-      usuallyRequired: false,
-      isMovement: false,
-      v1Only: false,
-      routing: false,
-      practiceSection: 'Spirit Temple',
-      notes: null,
-    },
-    {
-      id: 'B1F254D7',
-      trickName: 'Knuckles Fights?',
-      trickUrl: null,
-      isFundamental: true,
-      usuallyRequired: false,
-      isMovement: false,
-      v1Only: false,
-      routing: false,
-      practiceSection: 'Spirit Temple',
-      notes: null,
-    },
-    {
-      id: 'FF124B82',
-      trickName: 'Hoverslide BK Skip',
-      trickUrl: 'https://www.youtube.com/watch?v=EuiJlqaV76c',
-      isFundamental: false,
-      usuallyRequired: false,
-      isMovement: false,
-      v1Only: false,
-      routing: false,
-      practiceSection: 'Spirit Temple',
-      notes: null,
-    },
-    {
-      id: '42513477',
-      trickName: 'Spirit Entrance Block Skip',
-      trickUrl: 'https://www.youtube.com/watch?v=QisEMB_uNRA',
-      isFundamental: false,
-      usuallyRequired: false,
-      isMovement: false,
-      v1Only: false,
-      routing: false,
-      practiceSection: 'Spirit Temple',
-      notes: null,
-    },
-    {
-      id: 'DAC8501B',
-      trickName: 'Spirit SoT Skull',
-      trickUrl: 'https://youtu.be/eeT4BRTQN8Y?t=98',
-      isFundamental: false,
-      usuallyRequired: false,
-      isMovement: false,
-      v1Only: false,
-      routing: false,
-      practiceSection: 'Spirit Temple',
-      notes: null,
-    },
-    {
-      id: 'D7E0B9BE',
-      trickName: 'Spirit Child Side as Adult',
-      trickUrl: 'https://youtu.be/eeT4BRTQN8Y?t=138',
-      isFundamental: false,
-      usuallyRequired: false,
-      isMovement: false,
-      v1Only: false,
-      routing: false,
-      practiceSection: 'Spirit Temple',
-      notes: null,
-    },
-    {
-      id: '3C1E3767',
-      trickName: 'Spirit Child Slide 2nd crawlspace skip',
-      trickUrl: 'https://youtu.be/eeT4BRTQN8Y?t=188',
-      isFundamental: false,
-      usuallyRequired: false,
-      isMovement: false,
-      v1Only: false,
-      routing: false,
-      practiceSection: 'Spirit Temple',
-      notes: null,
-    },
-    {
-      id: '27D58B62',
-      trickName: 'Nabooru Knuckle Route',
-      trickUrl: 'https://www.youtube.com/watch?v=YJrz_2o_h_Y',
-      isFundamental: false,
-      usuallyRequired: false,
-      isMovement: false,
-      v1Only: false,
-      routing: false,
-      practiceSection: 'Spirit Temple',
-      notes: null,
-    },
-    {
-      id: '58D39228',
-      trickName: 'Spirit Map',
-      trickUrl: 'https://www.youtube.com/watch?v=lagMotf1w7o',
-      isFundamental: false,
-      usuallyRequired: false,
-      isMovement: false,
-      v1Only: false,
-      routing: false,
-      practiceSection: 'Spirit Temple',
-      notes: null,
-    },
-    {
-      id: '33028619',
-      trickName: 'All 4 Jabu Skulls Route',
-      trickUrl:
-        'https://www.youtube.com/watch?v=0m75oSmJvL4&list=PLis_W6MEPku3PAobg6W1zP0-N8j7e-l08&index=3&t=0s',
-      isFundamental: false,
-      usuallyRequired: false,
-      isMovement: false,
-      v1Only: false,
-      routing: false,
-      practiceSection: 'Jabu Jabu',
-      notes: null,
-    },
-    {
-      id: '933844BC',
-      trickName: 'Jabu Jabu as Adult',
-      trickUrl: 'https://www.youtube.com/watch?v=pPqKkCcBtk0',
-      isFundamental: false,
-      usuallyRequired: false,
-      isMovement: false,
-      v1Only: false,
-      routing: false,
-      practiceSection: 'Jabu Jabu',
-      notes: null,
-    },
-    {
-      id: '81D84A88',
-      trickName: 'Hover to big octo',
-      trickUrl: null,
-      isFundamental: false,
-      usuallyRequired: false,
-      isMovement: false,
-      v1Only: false,
-      routing: false,
-      practiceSection: 'Jabu Jabu',
-      notes: null,
-    },
-    {
-      id: 'D14D73DC',
-      trickName: '5 Magic Beans Route',
-      trickUrl: 'https://www.youtube.com/watch?v=cTlD9I7Wpyk',
-      isFundamental: false,
-      usuallyRequired: false,
-      isMovement: false,
-      v1Only: false,
-      routing: false,
-      practiceSection: 'Generic Overworld',
-      notes: null,
-    },
-    {
-      id: '3F3D3A81',
-      trickName: '1 Skull from each Child Dungeon Route',
-      trickUrl: null,
-      isFundamental: false,
-      usuallyRequired: false,
-      isMovement: false,
-      v1Only: false,
-      routing: false,
-      practiceSection: 'Generic Overworld',
-      notes: null,
-    },
-    {
-      id: '69AB0D65',
-      trickName: '3 Swords Route',
-      trickUrl: 'https://youtu.be/W9prYyTaZh8',
-      isFundamental: false,
-      usuallyRequired: false,
-      isMovement: false,
-      v1Only: false,
-      routing: false,
-      practiceSection: 'Generic Overworld',
-      notes: null,
-    },
-    {
-      id: '1660C305',
-      trickName: 'Defeat Skull Kid ISG',
-      trickUrl: 'https://youtu.be/i7htaVwLJm4',
-      isFundamental: false,
-      usuallyRequired: false,
-      isMovement: false,
-      v1Only: false,
-      routing: false,
-      practiceSection: 'Generic Overworld',
-      notes: null,
-    },
-    {
-      id: '2AC9A6C2',
-      trickName: 'Bombchu Bowling Strategy',
-      trickUrl: 'https://youtu.be/asTd9OOAGq0',
-      isFundamental: false,
-      usuallyRequired: false,
-      isMovement: false,
-      v1Only: false,
-      routing: false,
-      practiceSection: 'Generic Overworld',
-      notes: null,
-    },
-    {
-      id: '5FCDBB54',
-      trickName: 'Equip Swap',
-      trickUrl: 'https://youtu.be/Ul2ipeV5npQ?t=14',
-      isFundamental: true,
-      usuallyRequired: true,
-      isMovement: false,
-      v1Only: false,
-      routing: false,
-      practiceSection: 'Generic Overworld',
-      notes: null,
-    },
-    {
-      id: '81D84A88',
-      trickName: 'Lost Woods Skull Hookshot',
-      trickUrl: 'https://www.youtube.com/watch?v=IEcRdmaoqAM',
-      isFundamental: true,
-      usuallyRequired: true,
-      isMovement: false,
-      v1Only: false,
-      routing: false,
-      practiceSection: 'Generic Overworld',
-      notes: null,
-    },
-    {
-      id: 'BC67E767',
-      trickName: 'Lost Woods Skulltchula Duping',
-      trickUrl: 'https://www.youtube.com/watch?v=XVE7xiQHxck&t=14s',
-      isFundamental: false,
-      usuallyRequired: false,
-      isMovement: false,
-      v1Only: false,
-      routing: false,
-      practiceSection: 'Generic Overworld',
-      notes: null,
-    },
-    {
-      id: '7B3EABC2',
-      trickName: 'Adult GTG Keys Route',
-      trickUrl: 'https://www.youtube.com/watch?v=At-c_kJPcsA',
-      isFundamental: false,
-      usuallyRequired: false,
-      isMovement: false,
-      v1Only: false,
-      routing: false,
-      practiceSection: 'GTG',
-      notes: null,
-    },
-    {
-      id: 'A596CE56',
-      trickName: 'Weirdshot to Ice Arrows',
-      trickUrl:
-        'https://www.youtube.com/watch?v=MARrDb8KO9g&list=PLis_W6MEPku2SDFOH9jpmNMgdFFjZ9JBr&index=4',
-      isFundamental: false,
-      usuallyRequired: false,
-      isMovement: false,
-      v1Only: false,
-      routing: false,
-      practiceSection: 'GTG',
-      notes: null,
-    },
-    {
-      id: '2BD6E687',
-      trickName: 'Child Ice Arrows With Vine Clip',
-      trickUrl: 'https://www.youtube.com/watch?v=kNMCivcY2Nc',
-      isFundamental: false,
-      usuallyRequired: false,
-      isMovement: false,
-      v1Only: false,
-      routing: false,
-      practiceSection: 'GTG',
-      notes: null,
-    },
-    {
-      id: '2C5B71A5',
-      trickName: 'Child GTG Keys Route',
-      trickUrl: 'https://www.youtube.com/watch?v=T9CBkb_MR-I',
-      isFundamental: false,
-      usuallyRequired: false,
-      isMovement: false,
-      v1Only: false,
-      routing: false,
-      practiceSection: 'GTG',
-      notes: null,
-    },
-    {
-      id: '5EE45339',
-      trickName: 'Clip into GTG',
-      trickUrl: 'https://www.youtube.com/watch?v=bcCYAhsI8bY',
-      isFundamental: false,
-      usuallyRequired: false,
-      isMovement: false,
-      v1Only: false,
-      routing: false,
-      practiceSection: 'GTG',
-      notes: null,
-    },
-    {
-      id: '3126A906',
-      trickName: 'BotW 3 Skulls Route',
-      trickUrl:
-        'https://www.youtube.com/watch?v=gZO1l-tNqaI&feature=youtu.be&list=PLXd60MhAT2ss7gDa9LXU7p54rGBpmXtwT&t=75',
-      isFundamental: false,
-      usuallyRequired: false,
-      isMovement: false,
-      v1Only: false,
-      routing: false,
-      practiceSection: 'Bottom of the Well',
-      notes: null,
-    },
-    {
-      id: '7A34B2AD',
-      trickName: 'Actor Glitch to Deadhand',
-      trickUrl: 'https://www.youtube.com/watch?v=tTAtNXyR3wg',
-      isFundamental: false,
-      usuallyRequired: false,
-      isMovement: false,
-      v1Only: false,
-      routing: false,
-      practiceSection: 'Bottom of the Well',
-      notes: null,
-    },
-    {
-      id: '061F7FDE',
-      trickName: 'Early Chus With Ocarina',
-      trickUrl: 'https://youtu.be/FCNTgM0qIYI?t=1705',
-      isFundamental: true,
-      usuallyRequired: true,
-      isMovement: false,
-      v1Only: false,
-      routing: false,
-      practiceSection: 'Bottom of the Well',
-      notes: null,
-    },
-    {
-      id: '06B33728',
-      trickName: 'Early Chus Blank A',
-      trickUrl: 'https://youtu.be/Fz-Ekjg7jmA?t=1140',
-      isFundamental: true,
-      usuallyRequired: true,
-      isMovement: false,
-      v1Only: true,
-      routing: false,
-      practiceSection: 'Bottom of the Well',
-      notes: null,
-    },
-    {
-      id: '622C73E3',
-      trickName: 'Map in BotW',
-      trickUrl: null,
-      isFundamental: false,
-      usuallyRequired: false,
-      isMovement: false,
-      v1Only: false,
-      routing: false,
-      practiceSection: 'Bottom of the Well',
-      notes: null,
-    },
-    {
-      id: '3E3DECA3',
-      trickName: 'Graveyard Child SS Teleport',
-      trickUrl: 'https://www.youtube.com/watch?v=JeFR4uZeD2A',
-      isFundamental: false,
-      usuallyRequired: false,
-      isMovement: false,
-      v1Only: false,
-      routing: false,
-      practiceSection: 'Kak',
-      notes: null,
-    },
-    {
-      id: 'F09CEAC1',
-      trickName: 'Kak HP Dupe',
-      trickUrl: 'https://youtu.be/ZI_l1UTC9dM',
-      isFundamental: false,
-      usuallyRequired: false,
-      isMovement: false,
-      v1Only: false,
-      routing: false,
-      practiceSection: 'Kak',
-      notes: null,
-    },
-    {
-      id: '5CC02919',
-      trickName: 'Bottle on B (Outside Kak)',
-      trickUrl: null,
-      isFundamental: true,
-      usuallyRequired: true,
-      isMovement: false,
-      v1Only: false,
-      routing: false,
-      practiceSection: 'Kak',
-      notes: null,
-    },
-    {
-      id: 'A849D697',
-      trickName: 'RBA',
-      trickUrl: null,
-      isFundamental: true,
-      usuallyRequired: true,
-      isMovement: false,
-      v1Only: false,
-      routing: false,
-      practiceSection: 'Kak',
-      notes: 'RBA Chart: https://imgur.com/a/LgJdU',
-    },
-    {
-      id: '4689699D',
-      trickName: 'Dampe Void',
-      trickUrl: 'https://youtu.be/hBtBZILGepc',
-      isFundamental: true,
-      usuallyRequired: true,
-      isMovement: false,
-      v1Only: false,
-      routing: false,
-      practiceSection: 'Kak',
-      notes: null,
-    },
-    {
-      id: '32F255A9',
-      trickName: 'Hover boots into zoras domain',
-      trickUrl: 'https://youtu.be/B0kbnE3AmYc?t=2',
-      isFundamental: true,
-      usuallyRequired: true,
-      isMovement: false,
-      v1Only: false,
-      routing: false,
-      practiceSection: "Zora's Domain",
-      notes: null,
-    },
-    {
-      id: '4674313E',
-      trickName: 'Mega sidehop into zoras domain',
-      trickUrl: 'https://www.youtube.com/watch?v=9dDQUvB60Jk',
-      isFundamental: true,
-      usuallyRequired: false,
-      isMovement: false,
-      v1Only: false,
-      routing: false,
-      practiceSection: "Zora's Domain",
-      notes: null,
-    },
-    {
-      id: 'B0CA9C91',
-      trickName: 'Unfreeze King Zora',
-      trickUrl: 'https://www.youtube.com/watch?v=73O5FHcmdFE',
-      isFundamental: false,
-      usuallyRequired: false,
-      isMovement: false,
-      v1Only: false,
-      routing: false,
-      practiceSection: "Zora's Domain",
-      notes: null,
-    },
-    {
-      id: 'B5BDCB36',
-      trickName: 'Acute angle clip past King Zora',
-      trickUrl: 'https://youtu.be/5kJ3GG6fxCk',
-      isFundamental: true,
-      usuallyRequired: false,
-      isMovement: false,
-      v1Only: false,
-      routing: false,
-      practiceSection: "Zora's Domain",
-      notes: null,
-    },
-    {
-      id: 'E92BFD8B',
-      trickName: 'Child Chicken into Zora Domain',
-      trickUrl: 'https://youtu.be/o-Q29_V70II',
-      isFundamental: true,
-      usuallyRequired: false,
-      isMovement: false,
-      v1Only: false,
-      routing: false,
-      practiceSection: "Zora's Domain",
-      notes: null,
-    },
-    {
-      id: '089F7B22',
-      trickName: 'Child Hover to Skull',
-      trickUrl: 'https://www.youtube.com/watch?v=bJvJraCKCl0',
-      isFundamental: false,
-      usuallyRequired: false,
-      isMovement: false,
-      v1Only: false,
-      routing: false,
-      practiceSection: "Zora's Domain",
-      notes: null,
-    },
-    {
-      id: '366B4776',
-      trickName: 'Child Clip past King Zora',
-      trickUrl: 'https://www.youtube.com/watch?v=bQ-rmFcWcGc',
-      isFundamental: false,
-      usuallyRequired: false,
-      isMovement: false,
-      v1Only: false,
-      routing: false,
-      practiceSection: "Zora's Domain",
-      notes: null,
-    },
-    {
-      id: 'E54BB163',
-      trickName: 'Red Ice Clip To HP',
-      trickUrl: 'https://www.youtube.com/watch?v=l7x85j_i2ts',
-      isFundamental: false,
-      usuallyRequired: false,
-      isMovement: false,
-      v1Only: false,
-      routing: false,
-      practiceSection: 'Ice Cavern',
-      notes: null,
-    },
-    {
-      id: '0F804DD9',
-      trickName: 'Ledge Clip to HP and Compass',
-      trickUrl: 'https://youtu.be/a7jXaIcQes4',
-      isFundamental: false,
-      usuallyRequired: false,
-      isMovement: false,
-      v1Only: false,
-      routing: false,
-      practiceSection: 'Ice Cavern',
-      notes: null,
-    },
-    {
-      id: '827B2011',
-      trickName: 'Bomb Push Red Ice Skip',
-      trickUrl: 'https://www.youtube.com/watch?v=bThcEhgR7oM',
-      isFundamental: false,
-      usuallyRequired: false,
-      isMovement: false,
-      v1Only: false,
-      routing: false,
-      practiceSection: 'Ice Cavern',
-      notes: null,
-    },
-    {
-      id: '95759E64',
-      trickName: 'Serenade CS Skip',
-      trickUrl: 'https://youtu.be/Z5SGg34uD14?t=22',
-      isFundamental: true,
-      usuallyRequired: false,
-      isMovement: false,
-      v1Only: false,
-      routing: false,
-      practiceSection: 'Ice Cavern',
-      notes: null,
-    },
-    {
-      id: '6A60092E',
-      trickName: 'Get Compass Clips/Route',
-      trickUrl: 'https://youtu.be/a7jXaIcQes4',
-      isFundamental: true,
-      usuallyRequired: false,
-      isMovement: false,
-      v1Only: false,
-      routing: false,
-      practiceSection: 'Ice Cavern',
-      notes: null,
-    },
-    {
-      id: '13E1BC06',
-      trickName: 'Adult Deku Skulls Route',
-      trickUrl: 'https://www.youtube.com/watch?v=boxaS85Yxsk',
-      isFundamental: false,
-      usuallyRequired: false,
-      isMovement: false,
-      v1Only: false,
-      routing: false,
-      practiceSection: 'Deku Tree',
-      notes: null,
-    },
-    {
-      id: 'B669724E',
-      trickName: 'Adult Megajump to Deku Tree',
-      trickUrl: 'https://youtu.be/m6xPbz5m8hk',
-      isFundamental: false,
-      usuallyRequired: false,
-      isMovement: false,
-      v1Only: false,
-      routing: false,
-      practiceSection: 'Deku Tree',
-      notes: null,
-    },
-    {
-      id: '7958F88E',
-      trickName: 'Adult 231 Skip (Bombs)',
-      trickUrl: 'https://www.youtube.com/watch?v=4uZez2Oilec',
-      isFundamental: false,
-      usuallyRequired: false,
-      isMovement: false,
-      v1Only: false,
-      routing: false,
-      practiceSection: 'Deku Tree',
-      notes: null,
-    },
-    {
-      id: '2B40FB6F',
-      trickName: 'Adult 231 Skip (No Bombs)',
-      trickUrl: 'https://youtu.be/eGKF8dPMaSY',
-      isFundamental: false,
-      usuallyRequired: false,
-      isMovement: false,
-      v1Only: false,
-      routing: false,
-      practiceSection: 'Deku Tree',
-      notes: null,
-    },
-    {
-      id: 'F4B0B104',
-      trickName: 'DMC HP With Hovers',
-      trickUrl: 'https://www.youtube.com/watch?v=gS-0qzNoRaU',
-      isFundamental: false,
-      usuallyRequired: false,
-      isMovement: false,
-      v1Only: false,
-      routing: false,
-      practiceSection: 'Death Mountain',
-      notes: null,
-    },
-    {
-      id: '4657BEE7',
-      trickName: 'Bolero CS Skip',
-      trickUrl: 'https://youtu.be/CtZhmekDaNw',
-      isFundamental: true,
-      usuallyRequired: true,
-      isMovement: false,
-      v1Only: false,
-      routing: false,
-      practiceSection: 'Death Mountain',
-      notes: null,
-    },
-    {
-      id: '62527303',
-      trickName: 'Megaflip to double magic',
-      trickUrl: 'https://youtu.be/pkc6mm_RuPc',
-      isFundamental: false,
-      usuallyRequired: false,
-      isMovement: false,
-      v1Only: false,
-      routing: false,
-      practiceSection: 'Death Mountain',
-      notes: null,
-    },
-    {
-      id: '153192BA',
-      trickName: 'Megaflip to bean patch',
-      trickUrl: 'https://youtu.be/ODnM5pfC2hY',
-      isFundamental: false,
-      usuallyRequired: false,
-      isMovement: false,
-      v1Only: false,
-      routing: false,
-      practiceSection: 'Death Mountain',
-      notes: null,
-    },
-    {
-      id: '39077DAC',
-      trickName: 'Goron Shop Clip',
-      trickUrl:
-        'https://www.youtube.com/watch?v=cszpYcPMIJI&list=PL1n_Y5KrDdmUu3R2WUXPEmKjsg1XEr84u&index=19&t=0s',
-      isFundamental: false,
-      usuallyRequired: false,
-      isMovement: false,
-      v1Only: false,
-      routing: false,
-      practiceSection: 'Death Mountain',
-      notes: null,
-    },
-    {
-      id: '20BB24C5',
-      trickName: "Child Mega sidehop into Darunia's Room",
-      trickUrl: 'https://youtu.be/tMSiqzY3E2o',
-      isFundamental: false,
-      usuallyRequired: false,
-      isMovement: false,
-      v1Only: false,
-      routing: false,
-      practiceSection: 'Death Mountain',
-      notes: null,
-    },
-    {
-      id: '58A35C3D',
-      trickName: "Child Clip into DMC From Darunia's Room",
-      trickUrl: 'https://youtu.be/w4roOf4eC2Q',
-      isFundamental: false,
-      usuallyRequired: false,
-      isMovement: false,
-      v1Only: false,
-      routing: false,
-      practiceSection: 'Death Mountain',
-      notes: null,
-    },
-    {
-      id: '60531191',
-      trickName: 'All 4 Lon Lon Skulls (Child No Boomerang)',
-      trickUrl: 'https://www.youtube.com/watch?v=X8qQGLpye54',
-      isFundamental: false,
-      usuallyRequired: false,
-      isMovement: false,
-      v1Only: false,
-      routing: false,
-      practiceSection: 'Lon Lon Ranch',
-      notes: null,
-    },
-    {
-      id: '2ECE35A9',
-      trickName: 'Steal Epona With Hover',
-      trickUrl: 'https://youtu.be/sFQHyKT2ACQ',
-      isFundamental: false,
-      usuallyRequired: false,
-      isMovement: false,
-      v1Only: false,
-      routing: false,
-      practiceSection: 'Lon Lon Ranch',
-      notes: null,
-    },
-    {
-      id: 'AE1ED4F9',
-      trickName: 'Lon Lon HP as Adult',
-      trickUrl: 'https://www.youtube.com/watch?v=9L4eEBTegXo',
-      isFundamental: false,
-      usuallyRequired: false,
-      isMovement: false,
-      v1Only: false,
-      routing: false,
-      practiceSection: 'Lon Lon Ranch',
-      notes: null,
-    },
-    {
-      id: '8C37492D',
-      trickName: 'Hover to Fire Arrows',
-      trickUrl:
-        'https://youtu.be/bwLtoVynX0k?list=PLis_W6MEPku3PAobg6W1zP0-N8j7e-l08&t=1417',
-      isFundamental: false,
-      usuallyRequired: false,
-      isMovement: false,
-      v1Only: false,
-      routing: false,
-      practiceSection: 'Lake Hylia',
-      notes: null,
-    },
-    {
-      id: 'A17F4252',
-      trickName: 'Hookshot Jump To Fire Arrows',
-      trickUrl:
-        'https://www.youtube.com/watch?v=fvSUL82reH0&list=PLis_W6MEPku2SDFOH9jpmNMgdFFjZ9JBr&index=3&t=0s',
-      isFundamental: false,
-      usuallyRequired: false,
-      isMovement: false,
-      v1Only: false,
-      routing: false,
-      practiceSection: 'Lake Hylia',
-      notes: null,
-    },
-    {
-      id: 'C34B7975',
-      trickName: 'Jumpslash to House Gold Skull',
-      trickUrl: 'https://www.youtube.com/watch?v=GKC4dHbDK3g',
-      isFundamental: false,
-      usuallyRequired: false,
-      isMovement: false,
-      v1Only: false,
-      routing: false,
-      practiceSection: 'Lake Hylia',
-      notes: null,
-    },
-    {
-      id: 'B1F50951',
-      trickName: "Megaflip into Zora's Domain from Hylia",
-      trickUrl: 'https://www.youtube.com/watch?v=tA0IXmII6e4',
-      isFundamental: false,
-      usuallyRequired: false,
-      isMovement: false,
-      v1Only: false,
-      routing: true,
-      practiceSection: 'Lake Hylia',
-      notes:
-        'Could use a better video which shows the ESS setup + what frames you can roll on for megaflip',
-    },
-    {
-      id: 'FA10A855',
-      trickName: 'Contortion Hover to Ganons Castle',
-      trickUrl: 'https://www.youtube.com/watch?v=L5PbgPlfT04',
-      isFundamental: true,
-      usuallyRequired: false,
-      isMovement: false,
-      v1Only: false,
-      routing: false,
-      practiceSection: "Ganon's Castle",
-      notes: null,
-    },
-    {
-      id: '1BFB5B64',
-      trickName: 'Fire Trial',
-      trickUrl: 'https://youtu.be/vStsV0Kew90?t=13',
-      isFundamental: false,
-      usuallyRequired: false,
-      isMovement: false,
-      v1Only: false,
-      routing: false,
-      practiceSection: "Ganon's Castle",
-      notes: null,
-    },
-    {
-      id: 'E67D7D5A',
-      trickName: 'Forest Trial',
-      trickUrl: 'https://www.youtube.com/watch?v=dfSJycHQiqE',
-      isFundamental: false,
-      usuallyRequired: false,
-      isMovement: false,
-      v1Only: false,
-      routing: false,
-      practiceSection: "Ganon's Castle",
-      notes: null,
-    },
-    {
-      id: 'FB51F631',
-      trickName: 'Clip into Light Trial',
-      trickUrl: 'https://youtu.be/V9L-ePeEhH0',
-      isFundamental: false,
-      usuallyRequired: false,
-      isMovement: false,
-      v1Only: false,
-      routing: false,
-      practiceSection: "Ganon's Castle",
-      notes: null,
-    },
-    {
-      id: '8263CAC2',
-      trickName: 'Shadow Trial',
-      trickUrl: 'https://youtu.be/sFQHyKT2ACQ',
-      isFundamental: false,
-      usuallyRequired: false,
-      isMovement: false,
-      v1Only: false,
-      routing: false,
-      practiceSection: "Ganon's Castle",
-      notes: null,
-    },
-    {
-      id: '9E248370',
-      trickName: 'Water Trial',
-      trickUrl: 'https://youtu.be/6g_ggyB-xRw',
-      isFundamental: false,
-      usuallyRequired: false,
-      isMovement: false,
-      v1Only: false,
-      routing: false,
-      practiceSection: "Ganon's Castle",
-      notes: null,
-    },
-    {
-      id: '7ED23152',
-      trickName: 'Clip into Double Defense',
-      trickUrl: 'https://www.youtube.com/watch?v=I3pwrLRSo8U',
-      isFundamental: false,
-      usuallyRequired: false,
-      isMovement: false,
-      v1Only: false,
-      routing: false,
-      practiceSection: "Ganon's Castle",
-      notes: null,
-    },
-    {
-      id: 'EBBD3E69',
-      trickName: 'Mega sidehop trials skip',
-      trickUrl: 'https://www.youtube.com/watch?v=Dw4y0B20nM4',
-      isFundamental: false,
-      usuallyRequired: false,
-      isMovement: false,
-      v1Only: false,
-      routing: false,
-      practiceSection: "Ganon's Castle",
-      notes: null,
-    },
-    {
-      id: 'A87720FA',
-      trickName: 'Hover Trials Skip',
-      trickUrl: 'https://www.youtube.com/watch?v=MQZk6-R3dZA',
-      isFundamental: false,
-      usuallyRequired: false,
-      isMovement: false,
-      v1Only: false,
-      routing: false,
-      practiceSection: "Ganon's Castle",
-      notes: null,
-    },
-    {
-      id: '452A0F7F',
-      trickName: 'Statue Trials skip',
-      trickUrl: 'https://youtu.be/AZGccdS6nkQ?t=2',
-      isFundamental: false,
-      usuallyRequired: false,
-      isMovement: false,
-      v1Only: false,
-      routing: false,
-      practiceSection: "Ganon's Castle",
-      notes: null,
-    },
-    {
-      id: '1AD2FFF3',
-      trickName: 'Double Knuckles Fight Strategy',
-      trickUrl: 'https://youtu.be/AgBByp1nCl4?t=14561',
-      isFundamental: false,
-      usuallyRequired: false,
-      isMovement: false,
-      v1Only: false,
-      routing: false,
-      practiceSection: "Ganon's Castle",
-      notes: null,
-    },
-  ];
-};
