+++ data-access-layer/repository-interfaces/category-repository.ts
@@ -1,0 +1,6 @@
+import { Category } from '../../domain-import-only/Category';
+
+export interface ICategoryRepository {
+  getCategoryByName(name: string): Promise<Category | null>;
+  getAllCategories(): Promise<Category[]>;
+}
