+++ data-access-layer/repository-interfaces/route-repository.ts
@@ -1,0 +1,6 @@
+import { Route } from '../../domain-import-only/Route';
+
+export interface IRouteRepository {
+  getRouteByName(name: string): Promise<Route | null>;
+  getAllRoutes(): Promise<Route[]>;
+}
